Struct aws_sdk_chime::client::fluent_builders::CreateProxySession

pub struct CreateProxySession<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateProxySession<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateProxySessionOutput, SdkError<CreateProxySessionError>> where
R::Policy: SmithyRetryPolicy<CreateProxySessionInputOperationOutputAlias, CreateProxySessionOutput, CreateProxySessionError, CreateProxySessionInputOperationRetryAlias>,

pub fn voice_connector_id(self, inp: impl Into<String>) -> Self

The Amazon Chime voice connector ID.

pub fn set_voice_connector_id(self, input: Option<String>) -> Self

pub fn participant_phone_numbers(self, inp: impl Into<String>) -> Self

Appends an item to ParticipantPhoneNumbers.

To override the contents of this collection use set_participant_phone_numbers.

The participant phone numbers.

pub fn set_participant_phone_numbers(self, input: Option<Vec<String>>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the proxy session.

pub fn set_name(self, input: Option<String>) -> Self

pub fn expiry_minutes(self, inp: i32) -> Self

The number of minutes allowed for the proxy session.

pub fn set_expiry_minutes(self, input: Option<i32>) -> Self

pub fn capabilities(self, inp: impl Into<Capability>) -> Self

Appends an item to Capabilities.

To override the contents of this collection use set_capabilities.

The proxy session capabilities.

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn number_selection_behavior(self, inp: NumberSelectionBehavior) -> Self

The preference for proxy phone number reuse, or stickiness, between the same participants across sessions.

pub fn set_number_selection_behavior(
self,
input: Option<NumberSelectionBehavior>
) -> Self

pub fn geo_match_level(self, inp: GeoMatchLevel) -> Self

The preference for matching the country or area code of the proxy phone number with that of the first participant.

pub fn set_geo_match_level(self, input: Option<GeoMatchLevel>) -> Self

pub fn geo_match_params(self, inp: GeoMatchParams) -> Self

The country and area code for the proxy phone number.

pub fn set_geo_match_params(self, input: Option<GeoMatchParams>) -> Self