Struct aws_sdk_chime::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for UCBuzzConsoleService.

This client allows ergonomic access to a UCBuzzConsoleService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_phone_numbers_with_voice_connector(
&self
) -> AssociatePhoneNumbersWithVoiceConnector<C, M, R>

pub fn associate_phone_numbers_with_voice_connector_group(
&self
) -> AssociatePhoneNumbersWithVoiceConnectorGroup<C, M, R>

pub fn associate_phone_number_with_user(
&self
) -> AssociatePhoneNumberWithUser<C, M, R>

pub fn associate_signin_delegate_groups_with_account(
&self
) -> AssociateSigninDelegateGroupsWithAccount<C, M, R>

pub fn batch_create_attendee(&self) -> BatchCreateAttendee<C, M, R>

pub fn batch_create_channel_membership(
&self
) -> BatchCreateChannelMembership<C, M, R>

pub fn batch_create_room_membership(&self) -> BatchCreateRoomMembership<C, M, R>

pub fn batch_delete_phone_number(&self) -> BatchDeletePhoneNumber<C, M, R>

pub fn batch_suspend_user(&self) -> BatchSuspendUser<C, M, R>

pub fn batch_unsuspend_user(&self) -> BatchUnsuspendUser<C, M, R>

pub fn batch_update_phone_number(&self) -> BatchUpdatePhoneNumber<C, M, R>

pub fn batch_update_user(&self) -> BatchUpdateUser<C, M, R>

pub fn create_account(&self) -> CreateAccount<C, M, R>

pub fn create_app_instance(&self) -> CreateAppInstance<C, M, R>

pub fn create_app_instance_admin(&self) -> CreateAppInstanceAdmin<C, M, R>

pub fn create_app_instance_user(&self) -> CreateAppInstanceUser<C, M, R>

pub fn create_attendee(&self) -> CreateAttendee<C, M, R>

pub fn create_bot(&self) -> CreateBot<C, M, R>

pub fn create_channel(&self) -> CreateChannel<C, M, R>

pub fn create_channel_ban(&self) -> CreateChannelBan<C, M, R>

pub fn create_channel_membership(&self) -> CreateChannelMembership<C, M, R>

pub fn create_channel_moderator(&self) -> CreateChannelModerator<C, M, R>

pub fn create_media_capture_pipeline(
&self
) -> CreateMediaCapturePipeline<C, M, R>

pub fn create_meeting(&self) -> CreateMeeting<C, M, R>

pub fn create_meeting_dial_out(&self) -> CreateMeetingDialOut<C, M, R>

pub fn create_meeting_with_attendees(
&self
) -> CreateMeetingWithAttendees<C, M, R>

pub fn create_phone_number_order(&self) -> CreatePhoneNumberOrder<C, M, R>

pub fn create_proxy_session(&self) -> CreateProxySession<C, M, R>

pub fn create_room(&self) -> CreateRoom<C, M, R>

pub fn create_room_membership(&self) -> CreateRoomMembership<C, M, R>

pub fn create_sip_media_application(&self) -> CreateSipMediaApplication<C, M, R>

pub fn create_sip_media_application_call(
&self
) -> CreateSipMediaApplicationCall<C, M, R>

pub fn create_sip_rule(&self) -> CreateSipRule<C, M, R>

pub fn create_user(&self) -> CreateUser<C, M, R>

pub fn create_voice_connector(&self) -> CreateVoiceConnector<C, M, R>

pub fn create_voice_connector_group(&self) -> CreateVoiceConnectorGroup<C, M, R>

pub fn delete_account(&self) -> DeleteAccount<C, M, R>

pub fn delete_app_instance(&self) -> DeleteAppInstance<C, M, R>

pub fn delete_app_instance_admin(&self) -> DeleteAppInstanceAdmin<C, M, R>

pub fn delete_app_instance_streaming_configurations(
&self
) -> DeleteAppInstanceStreamingConfigurations<C, M, R>

pub fn delete_app_instance_user(&self) -> DeleteAppInstanceUser<C, M, R>

pub fn delete_attendee(&self) -> DeleteAttendee<C, M, R>

pub fn delete_channel(&self) -> DeleteChannel<C, M, R>

pub fn delete_channel_ban(&self) -> DeleteChannelBan<C, M, R>

pub fn delete_channel_membership(&self) -> DeleteChannelMembership<C, M, R>

pub fn delete_channel_message(&self) -> DeleteChannelMessage<C, M, R>

pub fn delete_channel_moderator(&self) -> DeleteChannelModerator<C, M, R>

pub fn delete_events_configuration(&self) -> DeleteEventsConfiguration<C, M, R>

pub fn delete_media_capture_pipeline(
&self
) -> DeleteMediaCapturePipeline<C, M, R>

pub fn delete_meeting(&self) -> DeleteMeeting<C, M, R>

pub fn delete_phone_number(&self) -> DeletePhoneNumber<C, M, R>

pub fn delete_proxy_session(&self) -> DeleteProxySession<C, M, R>

pub fn delete_room(&self) -> DeleteRoom<C, M, R>

pub fn delete_room_membership(&self) -> DeleteRoomMembership<C, M, R>

pub fn delete_sip_media_application(&self) -> DeleteSipMediaApplication<C, M, R>

pub fn delete_sip_rule(&self) -> DeleteSipRule<C, M, R>

pub fn delete_voice_connector(&self) -> DeleteVoiceConnector<C, M, R>

pub fn delete_voice_connector_emergency_calling_configuration(
&self
) -> DeleteVoiceConnectorEmergencyCallingConfiguration<C, M, R>

pub fn delete_voice_connector_group(&self) -> DeleteVoiceConnectorGroup<C, M, R>

pub fn delete_voice_connector_origination(
&self
) -> DeleteVoiceConnectorOrigination<C, M, R>

pub fn delete_voice_connector_proxy(&self) -> DeleteVoiceConnectorProxy<C, M, R>

pub fn delete_voice_connector_streaming_configuration(
&self
) -> DeleteVoiceConnectorStreamingConfiguration<C, M, R>

pub fn delete_voice_connector_termination(
&self
) -> DeleteVoiceConnectorTermination<C, M, R>

pub fn delete_voice_connector_termination_credentials(
&self
) -> DeleteVoiceConnectorTerminationCredentials<C, M, R>

pub fn describe_app_instance(&self) -> DescribeAppInstance<C, M, R>

pub fn describe_app_instance_admin(&self) -> DescribeAppInstanceAdmin<C, M, R>

pub fn describe_app_instance_user(&self) -> DescribeAppInstanceUser<C, M, R>

pub fn describe_channel(&self) -> DescribeChannel<C, M, R>

pub fn describe_channel_ban(&self) -> DescribeChannelBan<C, M, R>

pub fn describe_channel_membership(&self) -> DescribeChannelMembership<C, M, R>

pub fn describe_channel_membership_for_app_instance_user(
&self
) -> DescribeChannelMembershipForAppInstanceUser<C, M, R>

pub fn describe_channel_moderated_by_app_instance_user(
&self
) -> DescribeChannelModeratedByAppInstanceUser<C, M, R>

pub fn describe_channel_moderator(&self) -> DescribeChannelModerator<C, M, R>

pub fn disassociate_phone_number_from_user(
&self
) -> DisassociatePhoneNumberFromUser<C, M, R>

pub fn disassociate_phone_numbers_from_voice_connector(
&self
) -> DisassociatePhoneNumbersFromVoiceConnector<C, M, R>

pub fn disassociate_phone_numbers_from_voice_connector_group(
&self
) -> DisassociatePhoneNumbersFromVoiceConnectorGroup<C, M, R>

pub fn disassociate_signin_delegate_groups_from_account(
&self
) -> DisassociateSigninDelegateGroupsFromAccount<C, M, R>

pub fn get_account(&self) -> GetAccount<C, M, R>

pub fn get_account_settings(&self) -> GetAccountSettings<C, M, R>

pub fn get_app_instance_retention_settings(
&self
) -> GetAppInstanceRetentionSettings<C, M, R>

pub fn get_app_instance_streaming_configurations(
&self
) -> GetAppInstanceStreamingConfigurations<C, M, R>

pub fn get_attendee(&self) -> GetAttendee<C, M, R>

pub fn get_bot(&self) -> GetBot<C, M, R>

pub fn get_channel_message(&self) -> GetChannelMessage<C, M, R>

pub fn get_events_configuration(&self) -> GetEventsConfiguration<C, M, R>

pub fn get_global_settings(&self) -> GetGlobalSettings<C, M, R>

pub fn get_media_capture_pipeline(&self) -> GetMediaCapturePipeline<C, M, R>

pub fn get_meeting(&self) -> GetMeeting<C, M, R>

pub fn get_messaging_session_endpoint(
&self
) -> GetMessagingSessionEndpoint<C, M, R>

pub fn get_phone_number(&self) -> GetPhoneNumber<C, M, R>

pub fn get_phone_number_order(&self) -> GetPhoneNumberOrder<C, M, R>

pub fn get_phone_number_settings(&self) -> GetPhoneNumberSettings<C, M, R>

pub fn get_proxy_session(&self) -> GetProxySession<C, M, R>

pub fn get_retention_settings(&self) -> GetRetentionSettings<C, M, R>

pub fn get_room(&self) -> GetRoom<C, M, R>

pub fn get_sip_media_application(&self) -> GetSipMediaApplication<C, M, R>

pub fn get_sip_media_application_logging_configuration(
&self
) -> GetSipMediaApplicationLoggingConfiguration<C, M, R>

pub fn get_sip_rule(&self) -> GetSipRule<C, M, R>

pub fn get_user(&self) -> GetUser<C, M, R>

pub fn get_user_settings(&self) -> GetUserSettings<C, M, R>

pub fn get_voice_connector(&self) -> GetVoiceConnector<C, M, R>

pub fn get_voice_connector_emergency_calling_configuration(
&self
) -> GetVoiceConnectorEmergencyCallingConfiguration<C, M, R>

pub fn get_voice_connector_group(&self) -> GetVoiceConnectorGroup<C, M, R>

pub fn get_voice_connector_logging_configuration(
&self
) -> GetVoiceConnectorLoggingConfiguration<C, M, R>

pub fn get_voice_connector_origination(
&self
) -> GetVoiceConnectorOrigination<C, M, R>

pub fn get_voice_connector_proxy(&self) -> GetVoiceConnectorProxy<C, M, R>

pub fn get_voice_connector_streaming_configuration(
&self
) -> GetVoiceConnectorStreamingConfiguration<C, M, R>

pub fn get_voice_connector_termination(
&self
) -> GetVoiceConnectorTermination<C, M, R>

pub fn get_voice_connector_termination_health(
&self
) -> GetVoiceConnectorTerminationHealth<C, M, R>

pub fn invite_users(&self) -> InviteUsers<C, M, R>

pub fn list_accounts(&self) -> ListAccounts<C, M, R>

pub fn list_app_instance_admins(&self) -> ListAppInstanceAdmins<C, M, R>

pub fn list_app_instances(&self) -> ListAppInstances<C, M, R>

pub fn list_app_instance_users(&self) -> ListAppInstanceUsers<C, M, R>

pub fn list_attendees(&self) -> ListAttendees<C, M, R>

pub fn list_attendee_tags(&self) -> ListAttendeeTags<C, M, R>

pub fn list_bots(&self) -> ListBots<C, M, R>

pub fn list_channel_bans(&self) -> ListChannelBans<C, M, R>

pub fn list_channel_memberships(&self) -> ListChannelMemberships<C, M, R>

pub fn list_channel_memberships_for_app_instance_user(
&self
) -> ListChannelMembershipsForAppInstanceUser<C, M, R>

pub fn list_channel_messages(&self) -> ListChannelMessages<C, M, R>

pub fn list_channel_moderators(&self) -> ListChannelModerators<C, M, R>

pub fn list_channels(&self) -> ListChannels<C, M, R>

pub fn list_channels_moderated_by_app_instance_user(
&self
) -> ListChannelsModeratedByAppInstanceUser<C, M, R>

pub fn list_media_capture_pipelines(&self) -> ListMediaCapturePipelines<C, M, R>

pub fn list_meetings(&self) -> ListMeetings<C, M, R>

pub fn list_meeting_tags(&self) -> ListMeetingTags<C, M, R>

pub fn list_phone_number_orders(&self) -> ListPhoneNumberOrders<C, M, R>

pub fn list_phone_numbers(&self) -> ListPhoneNumbers<C, M, R>

pub fn list_proxy_sessions(&self) -> ListProxySessions<C, M, R>

pub fn list_room_memberships(&self) -> ListRoomMemberships<C, M, R>

pub fn list_rooms(&self) -> ListRooms<C, M, R>

pub fn list_sip_media_applications(&self) -> ListSipMediaApplications<C, M, R>

pub fn list_sip_rules(&self) -> ListSipRules<C, M, R>

pub fn list_supported_phone_number_countries(
&self
) -> ListSupportedPhoneNumberCountries<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_users(&self) -> ListUsers<C, M, R>

pub fn list_voice_connector_groups(&self) -> ListVoiceConnectorGroups<C, M, R>

pub fn list_voice_connectors(&self) -> ListVoiceConnectors<C, M, R>

pub fn list_voice_connector_termination_credentials(
&self
) -> ListVoiceConnectorTerminationCredentials<C, M, R>

pub fn logout_user(&self) -> LogoutUser<C, M, R>

pub fn put_app_instance_retention_settings(
&self
) -> PutAppInstanceRetentionSettings<C, M, R>

pub fn put_app_instance_streaming_configurations(
&self
) -> PutAppInstanceStreamingConfigurations<C, M, R>

pub fn put_events_configuration(&self) -> PutEventsConfiguration<C, M, R>

pub fn put_retention_settings(&self) -> PutRetentionSettings<C, M, R>

pub fn put_sip_media_application_logging_configuration(
&self
) -> PutSipMediaApplicationLoggingConfiguration<C, M, R>

pub fn put_voice_connector_emergency_calling_configuration(
&self
) -> PutVoiceConnectorEmergencyCallingConfiguration<C, M, R>

pub fn put_voice_connector_logging_configuration(
&self
) -> PutVoiceConnectorLoggingConfiguration<C, M, R>

pub fn put_voice_connector_origination(
&self
) -> PutVoiceConnectorOrigination<C, M, R>

pub fn put_voice_connector_proxy(&self) -> PutVoiceConnectorProxy<C, M, R>

pub fn put_voice_connector_streaming_configuration(
&self
) -> PutVoiceConnectorStreamingConfiguration<C, M, R>

pub fn put_voice_connector_termination(
&self
) -> PutVoiceConnectorTermination<C, M, R>

pub fn put_voice_connector_termination_credentials(
&self
) -> PutVoiceConnectorTerminationCredentials<C, M, R>

pub fn redact_channel_message(&self) -> RedactChannelMessage<C, M, R>

pub fn redact_conversation_message(&self) -> RedactConversationMessage<C, M, R>

pub fn redact_room_message(&self) -> RedactRoomMessage<C, M, R>

pub fn regenerate_security_token(&self) -> RegenerateSecurityToken<C, M, R>

pub fn reset_personal_pin(&self) -> ResetPersonalPIN<C, M, R>

pub fn restore_phone_number(&self) -> RestorePhoneNumber<C, M, R>

pub fn search_available_phone_numbers(
&self
) -> SearchAvailablePhoneNumbers<C, M, R>

pub fn send_channel_message(&self) -> SendChannelMessage<C, M, R>

pub fn start_meeting_transcription(&self) -> StartMeetingTranscription<C, M, R>

pub fn stop_meeting_transcription(&self) -> StopMeetingTranscription<C, M, R>

pub fn tag_attendee(&self) -> TagAttendee<C, M, R>

pub fn tag_meeting(&self) -> TagMeeting<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_attendee(&self) -> UntagAttendee<C, M, R>

pub fn untag_meeting(&self) -> UntagMeeting<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_account(&self) -> UpdateAccount<C, M, R>

pub fn update_account_settings(&self) -> UpdateAccountSettings<C, M, R>

pub fn update_app_instance(&self) -> UpdateAppInstance<C, M, R>

pub fn update_app_instance_user(&self) -> UpdateAppInstanceUser<C, M, R>

pub fn update_bot(&self) -> UpdateBot<C, M, R>

pub fn update_channel(&self) -> UpdateChannel<C, M, R>

pub fn update_channel_message(&self) -> UpdateChannelMessage<C, M, R>

pub fn update_channel_read_marker(&self) -> UpdateChannelReadMarker<C, M, R>

pub fn update_global_settings(&self) -> UpdateGlobalSettings<C, M, R>

pub fn update_phone_number(&self) -> UpdatePhoneNumber<C, M, R>

pub fn update_phone_number_settings(&self) -> UpdatePhoneNumberSettings<C, M, R>

pub fn update_proxy_session(&self) -> UpdateProxySession<C, M, R>

pub fn update_room(&self) -> UpdateRoom<C, M, R>

pub fn update_room_membership(&self) -> UpdateRoomMembership<C, M, R>

pub fn update_sip_media_application(&self) -> UpdateSipMediaApplication<C, M, R>

pub fn update_sip_media_application_call(
&self
) -> UpdateSipMediaApplicationCall<C, M, R>

pub fn update_sip_rule(&self) -> UpdateSipRule<C, M, R>

pub fn update_user(&self) -> UpdateUser<C, M, R>

pub fn update_user_settings(&self) -> UpdateUserSettings<C, M, R>

pub fn update_voice_connector(&self) -> UpdateVoiceConnector<C, M, R>

pub fn update_voice_connector_group(&self) -> UpdateVoiceConnectorGroup<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self