Enum aws_sdk_chime::Error

#[non_exhaustive]
pub enum Error {
    AccessDeniedException(AccessDeniedException),
    BadRequestException(BadRequestException),
    ConflictException(ConflictException),
    ForbiddenException(ForbiddenException),
    NotFoundException(NotFoundException),
    ResourceLimitExceededException(ResourceLimitExceededException),
    ServiceFailureException(ServiceFailureException),
    ServiceUnavailableException(ServiceUnavailableException),
    ThrottledClientException(ThrottledClientException),
    UnauthorizedClientException(UnauthorizedClientException),
    UnprocessableEntityException(UnprocessableEntityException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

You don't have permissions to perform the requested operation.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

BadRequestException(BadRequestException)

The input parameters don't match the service's restrictions.

Tuple Fields of BadRequestException

0: BadRequestException

ConflictException(ConflictException)

The request could not be processed because of conflict in the current state of the resource.

Tuple Fields of ConflictException

0: ConflictException

ForbiddenException(ForbiddenException)

The client is permanently forbidden from making the request.

Tuple Fields of ForbiddenException

0: ForbiddenException

NotFoundException(NotFoundException)

One or more of the resources in the request does not exist in the system.

Tuple Fields of NotFoundException

0: NotFoundException

ResourceLimitExceededException(ResourceLimitExceededException)

The request exceeds the resource limit.

Tuple Fields of ResourceLimitExceededException

0: ResourceLimitExceededException

ServiceFailureException(ServiceFailureException)

The service encountered an unexpected error.

Tuple Fields of ServiceFailureException

0: ServiceFailureException

ServiceUnavailableException(ServiceUnavailableException)

The service is currently unavailable.

Tuple Fields of ServiceUnavailableException

0: ServiceUnavailableException

ThrottledClientException(ThrottledClientException)

The client exceeded its request rate limit.

Tuple Fields of ThrottledClientException

0: ThrottledClientException

UnauthorizedClientException(UnauthorizedClientException)

The client is not currently authorized to make the request.

Tuple Fields of UnauthorizedClientException

0: UnauthorizedClientException

UnprocessableEntityException(UnprocessableEntityException)

The request was well-formed but was unable to be followed due to semantic errors.

Tuple Fields of UnprocessableEntityException

0: UnprocessableEntityException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>