Module aws_sdk_chime::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

not_found_exception

See NotFoundException

resource_limit_exceeded_exception

See ResourceLimitExceededException

service_failure_exception

See ServiceFailureException

service_unavailable_exception

See ServiceUnavailableException

throttled_client_exception

See ThrottledClientException

unauthorized_client_exception

See UnauthorizedClientException

unprocessable_entity_exception

See UnprocessableEntityException

Structs

AccessDeniedException

You don't have permissions to perform the requested operation.

AssociatePhoneNumberWithUserError

AssociatePhoneNumbersWithVoiceConnectorError

AssociatePhoneNumbersWithVoiceConnectorGroupError

AssociateSigninDelegateGroupsWithAccountError

BadRequestException

The input parameters don't match the service's restrictions.

BatchCreateAttendeeError

BatchCreateChannelMembershipError

BatchCreateRoomMembershipError

BatchDeletePhoneNumberError

BatchSuspendUserError

BatchUnsuspendUserError

BatchUpdatePhoneNumberError

BatchUpdateUserError

ConflictException

The request could not be processed because of conflict in the current state of the

CreateAccountError

CreateAppInstanceAdminError

CreateAppInstanceError

CreateAppInstanceUserError

CreateAttendeeError

CreateBotError

CreateChannelBanError

CreateChannelError

CreateChannelMembershipError

CreateChannelModeratorError

CreateMediaCapturePipelineError

CreateMeetingDialOutError

CreateMeetingError

CreateMeetingWithAttendeesError

CreatePhoneNumberOrderError

CreateProxySessionError

CreateRoomError

CreateRoomMembershipError

CreateSipMediaApplicationCallError

CreateSipMediaApplicationError

CreateSipRuleError

CreateUserError

CreateVoiceConnectorError

CreateVoiceConnectorGroupError

DeleteAccountError

DeleteAppInstanceAdminError

DeleteAppInstanceError

DeleteAppInstanceStreamingConfigurationsError

DeleteAppInstanceUserError

DeleteAttendeeError

DeleteChannelBanError

DeleteChannelError

DeleteChannelMembershipError

DeleteChannelMessageError

DeleteChannelModeratorError

DeleteEventsConfigurationError

DeleteMediaCapturePipelineError

DeleteMeetingError

DeletePhoneNumberError

DeleteProxySessionError

DeleteRoomError

DeleteRoomMembershipError

DeleteSipMediaApplicationError

DeleteSipRuleError

DeleteVoiceConnectorEmergencyCallingConfigurationError

DeleteVoiceConnectorError

DeleteVoiceConnectorGroupError

DeleteVoiceConnectorOriginationError

DeleteVoiceConnectorProxyError

DeleteVoiceConnectorStreamingConfigurationError

DeleteVoiceConnectorTerminationCredentialsError

DeleteVoiceConnectorTerminationError

DescribeAppInstanceAdminError

DescribeAppInstanceError

DescribeAppInstanceUserError

DescribeChannelBanError

DescribeChannelError

DescribeChannelMembershipError

DescribeChannelMembershipForAppInstanceUserError

DescribeChannelModeratedByAppInstanceUserError

DescribeChannelModeratorError

DisassociatePhoneNumberFromUserError

DisassociatePhoneNumbersFromVoiceConnectorError

DisassociatePhoneNumbersFromVoiceConnectorGroupError

DisassociateSigninDelegateGroupsFromAccountError

ForbiddenException

The client is permanently forbidden from making the request.

GetAccountError

GetAccountSettingsError

GetAppInstanceRetentionSettingsError

GetAppInstanceStreamingConfigurationsError

GetAttendeeError

GetBotError

GetChannelMessageError

GetEventsConfigurationError

GetGlobalSettingsError

GetMediaCapturePipelineError

GetMeetingError

GetMessagingSessionEndpointError

GetPhoneNumberError

GetPhoneNumberOrderError

GetPhoneNumberSettingsError

GetProxySessionError

GetRetentionSettingsError

GetRoomError

GetSipMediaApplicationError

GetSipMediaApplicationLoggingConfigurationError

GetSipRuleError

GetUserError

GetUserSettingsError

GetVoiceConnectorEmergencyCallingConfigurationError

GetVoiceConnectorError

GetVoiceConnectorGroupError

GetVoiceConnectorLoggingConfigurationError

GetVoiceConnectorOriginationError

GetVoiceConnectorProxyError

GetVoiceConnectorStreamingConfigurationError

GetVoiceConnectorTerminationError

GetVoiceConnectorTerminationHealthError

InviteUsersError

ListAccountsError

ListAppInstanceAdminsError

ListAppInstanceUsersError

ListAppInstancesError

ListAttendeeTagsError

ListAttendeesError

ListBotsError

ListChannelBansError

ListChannelMembershipsError

ListChannelMembershipsForAppInstanceUserError

ListChannelMessagesError

ListChannelModeratorsError

ListChannelsError

ListChannelsModeratedByAppInstanceUserError

ListMediaCapturePipelinesError

ListMeetingTagsError

ListMeetingsError

ListPhoneNumberOrdersError

ListPhoneNumbersError

ListProxySessionsError

ListRoomMembershipsError

ListRoomsError

ListSipMediaApplicationsError

ListSipRulesError

ListSupportedPhoneNumberCountriesError

ListTagsForResourceError

ListUsersError

ListVoiceConnectorGroupsError

ListVoiceConnectorTerminationCredentialsError

ListVoiceConnectorsError

LogoutUserError

NotFoundException

One or more of the resources in the request does not exist in the system.

PutAppInstanceRetentionSettingsError

PutAppInstanceStreamingConfigurationsError

PutEventsConfigurationError

PutRetentionSettingsError

PutSipMediaApplicationLoggingConfigurationError

PutVoiceConnectorEmergencyCallingConfigurationError

PutVoiceConnectorLoggingConfigurationError

PutVoiceConnectorOriginationError

PutVoiceConnectorProxyError

PutVoiceConnectorStreamingConfigurationError

PutVoiceConnectorTerminationCredentialsError

PutVoiceConnectorTerminationError

RedactChannelMessageError

RedactConversationMessageError

RedactRoomMessageError

RegenerateSecurityTokenError

ResetPersonalPINError

ResourceLimitExceededException

The request exceeds the resource limit.

RestorePhoneNumberError

SearchAvailablePhoneNumbersError

SendChannelMessageError

ServiceFailureException

The service encountered an unexpected error.

ServiceUnavailableException

The service is currently unavailable.

StartMeetingTranscriptionError

StopMeetingTranscriptionError

TagAttendeeError

TagMeetingError

TagResourceError

ThrottledClientException

The client exceeded its request rate limit.

UnauthorizedClientException

The client is not currently authorized to make the request.

UnprocessableEntityException

The request was well-formed but was unable to be followed due to semantic errors.

UntagAttendeeError

UntagMeetingError

UntagResourceError

UpdateAccountError

UpdateAccountSettingsError

UpdateAppInstanceError

UpdateAppInstanceUserError

UpdateBotError

UpdateChannelError

UpdateChannelMessageError

UpdateChannelReadMarkerError

UpdateGlobalSettingsError

UpdatePhoneNumberError

UpdatePhoneNumberSettingsError

UpdateProxySessionError

UpdateRoomError

UpdateRoomMembershipError

UpdateSipMediaApplicationCallError

UpdateSipMediaApplicationError

UpdateSipRuleError

UpdateUserError

UpdateUserSettingsError

UpdateVoiceConnectorError

UpdateVoiceConnectorGroupError

Enums

AssociatePhoneNumberWithUserErrorKind

AssociatePhoneNumbersWithVoiceConnectorErrorKind

AssociatePhoneNumbersWithVoiceConnectorGroupErrorKind

AssociateSigninDelegateGroupsWithAccountErrorKind

BatchCreateAttendeeErrorKind

BatchCreateChannelMembershipErrorKind

BatchCreateRoomMembershipErrorKind

BatchDeletePhoneNumberErrorKind

BatchSuspendUserErrorKind

BatchUnsuspendUserErrorKind

BatchUpdatePhoneNumberErrorKind

BatchUpdateUserErrorKind

CreateAccountErrorKind

CreateAppInstanceAdminErrorKind

CreateAppInstanceErrorKind

CreateAppInstanceUserErrorKind

CreateAttendeeErrorKind

CreateBotErrorKind

CreateChannelBanErrorKind

CreateChannelErrorKind

CreateChannelMembershipErrorKind

CreateChannelModeratorErrorKind

CreateMediaCapturePipelineErrorKind

CreateMeetingDialOutErrorKind

CreateMeetingErrorKind

CreateMeetingWithAttendeesErrorKind

CreatePhoneNumberOrderErrorKind

CreateProxySessionErrorKind

CreateRoomErrorKind

CreateRoomMembershipErrorKind

CreateSipMediaApplicationCallErrorKind

CreateSipMediaApplicationErrorKind

CreateSipRuleErrorKind

CreateUserErrorKind

CreateVoiceConnectorErrorKind

CreateVoiceConnectorGroupErrorKind

DeleteAccountErrorKind

DeleteAppInstanceAdminErrorKind

DeleteAppInstanceErrorKind

DeleteAppInstanceStreamingConfigurationsErrorKind

DeleteAppInstanceUserErrorKind

DeleteAttendeeErrorKind

DeleteChannelBanErrorKind

DeleteChannelErrorKind

DeleteChannelMembershipErrorKind

DeleteChannelMessageErrorKind

DeleteChannelModeratorErrorKind

DeleteEventsConfigurationErrorKind

DeleteMediaCapturePipelineErrorKind

DeleteMeetingErrorKind

DeletePhoneNumberErrorKind

DeleteProxySessionErrorKind

DeleteRoomErrorKind

DeleteRoomMembershipErrorKind

DeleteSipMediaApplicationErrorKind

DeleteSipRuleErrorKind

DeleteVoiceConnectorEmergencyCallingConfigurationErrorKind

DeleteVoiceConnectorErrorKind

DeleteVoiceConnectorGroupErrorKind

DeleteVoiceConnectorOriginationErrorKind

DeleteVoiceConnectorProxyErrorKind

DeleteVoiceConnectorStreamingConfigurationErrorKind

DeleteVoiceConnectorTerminationCredentialsErrorKind

DeleteVoiceConnectorTerminationErrorKind

DescribeAppInstanceAdminErrorKind

DescribeAppInstanceErrorKind

DescribeAppInstanceUserErrorKind

DescribeChannelBanErrorKind

DescribeChannelErrorKind

DescribeChannelMembershipErrorKind

DescribeChannelMembershipForAppInstanceUserErrorKind

DescribeChannelModeratedByAppInstanceUserErrorKind

DescribeChannelModeratorErrorKind

DisassociatePhoneNumberFromUserErrorKind

DisassociatePhoneNumbersFromVoiceConnectorErrorKind

DisassociatePhoneNumbersFromVoiceConnectorGroupErrorKind

DisassociateSigninDelegateGroupsFromAccountErrorKind

GetAccountErrorKind

GetAccountSettingsErrorKind

GetAppInstanceRetentionSettingsErrorKind

GetAppInstanceStreamingConfigurationsErrorKind

GetAttendeeErrorKind

GetBotErrorKind

GetChannelMessageErrorKind

GetEventsConfigurationErrorKind

GetGlobalSettingsErrorKind

GetMediaCapturePipelineErrorKind

GetMeetingErrorKind

GetMessagingSessionEndpointErrorKind

GetPhoneNumberErrorKind

GetPhoneNumberOrderErrorKind

GetPhoneNumberSettingsErrorKind

GetProxySessionErrorKind

GetRetentionSettingsErrorKind

GetRoomErrorKind

GetSipMediaApplicationErrorKind

GetSipMediaApplicationLoggingConfigurationErrorKind

GetSipRuleErrorKind

GetUserErrorKind

GetUserSettingsErrorKind

GetVoiceConnectorEmergencyCallingConfigurationErrorKind

GetVoiceConnectorErrorKind

GetVoiceConnectorGroupErrorKind

GetVoiceConnectorLoggingConfigurationErrorKind

GetVoiceConnectorOriginationErrorKind

GetVoiceConnectorProxyErrorKind

GetVoiceConnectorStreamingConfigurationErrorKind

GetVoiceConnectorTerminationErrorKind

GetVoiceConnectorTerminationHealthErrorKind

InviteUsersErrorKind

ListAccountsErrorKind

ListAppInstanceAdminsErrorKind

ListAppInstanceUsersErrorKind

ListAppInstancesErrorKind

ListAttendeeTagsErrorKind

ListAttendeesErrorKind

ListBotsErrorKind

ListChannelBansErrorKind

ListChannelMembershipsErrorKind

ListChannelMembershipsForAppInstanceUserErrorKind

ListChannelMessagesErrorKind

ListChannelModeratorsErrorKind

ListChannelsErrorKind

ListChannelsModeratedByAppInstanceUserErrorKind

ListMediaCapturePipelinesErrorKind

ListMeetingTagsErrorKind

ListMeetingsErrorKind

ListPhoneNumberOrdersErrorKind

ListPhoneNumbersErrorKind

ListProxySessionsErrorKind

ListRoomMembershipsErrorKind

ListRoomsErrorKind

ListSipMediaApplicationsErrorKind

ListSipRulesErrorKind

ListSupportedPhoneNumberCountriesErrorKind

ListTagsForResourceErrorKind

ListUsersErrorKind

ListVoiceConnectorGroupsErrorKind

ListVoiceConnectorTerminationCredentialsErrorKind

ListVoiceConnectorsErrorKind

LogoutUserErrorKind

PutAppInstanceRetentionSettingsErrorKind

PutAppInstanceStreamingConfigurationsErrorKind

PutEventsConfigurationErrorKind

PutRetentionSettingsErrorKind

PutSipMediaApplicationLoggingConfigurationErrorKind

PutVoiceConnectorEmergencyCallingConfigurationErrorKind

PutVoiceConnectorLoggingConfigurationErrorKind

PutVoiceConnectorOriginationErrorKind

PutVoiceConnectorProxyErrorKind

PutVoiceConnectorStreamingConfigurationErrorKind

PutVoiceConnectorTerminationCredentialsErrorKind

PutVoiceConnectorTerminationErrorKind

RedactChannelMessageErrorKind

RedactConversationMessageErrorKind

RedactRoomMessageErrorKind

RegenerateSecurityTokenErrorKind

ResetPersonalPINErrorKind

RestorePhoneNumberErrorKind

SearchAvailablePhoneNumbersErrorKind

SendChannelMessageErrorKind

StartMeetingTranscriptionErrorKind

StopMeetingTranscriptionErrorKind

TagAttendeeErrorKind

TagMeetingErrorKind

TagResourceErrorKind

UntagAttendeeErrorKind

UntagMeetingErrorKind

UntagResourceErrorKind

UpdateAccountErrorKind

UpdateAccountSettingsErrorKind

UpdateAppInstanceErrorKind

UpdateAppInstanceUserErrorKind

UpdateBotErrorKind

UpdateChannelErrorKind

UpdateChannelMessageErrorKind

UpdateChannelReadMarkerErrorKind

UpdateGlobalSettingsErrorKind

UpdatePhoneNumberErrorKind

UpdatePhoneNumberSettingsErrorKind

UpdateProxySessionErrorKind

UpdateRoomErrorKind

UpdateRoomMembershipErrorKind

UpdateSipMediaApplicationCallErrorKind

UpdateSipMediaApplicationErrorKind

UpdateSipRuleErrorKind

UpdateUserErrorKind

UpdateUserSettingsErrorKind

UpdateVoiceConnectorErrorKind

UpdateVoiceConnectorGroupErrorKind