Module aws_sdk_chime::model

Modules

account

See Account

account_settings

See AccountSettings

alexa_for_business_metadata

See AlexaForBusinessMetadata

app_instance

See AppInstance

app_instance_admin

See AppInstanceAdmin

app_instance_admin_summary

See AppInstanceAdminSummary

app_instance_retention_settings

See AppInstanceRetentionSettings

app_instance_streaming_configuration

See AppInstanceStreamingConfiguration

app_instance_summary

See AppInstanceSummary

app_instance_user

See AppInstanceUser

app_instance_user_membership_summary

See AppInstanceUserMembershipSummary

app_instance_user_summary

See AppInstanceUserSummary

attendee

See Attendee

batch_channel_memberships

See BatchChannelMemberships

batch_create_channel_membership_error

See BatchCreateChannelMembershipError

bot

See Bot

business_calling_settings

See BusinessCallingSettings

channel

See Channel

channel_ban

See ChannelBan

channel_ban_summary

See ChannelBanSummary

channel_membership

See ChannelMembership

channel_membership_for_app_instance_user_summary

See ChannelMembershipForAppInstanceUserSummary

channel_membership_summary

See ChannelMembershipSummary

channel_message

See ChannelMessage

channel_message_summary

See ChannelMessageSummary

channel_moderated_by_app_instance_user_summary

See ChannelModeratedByAppInstanceUserSummary

channel_moderator

See ChannelModerator

channel_moderator_summary

See ChannelModeratorSummary

channel_retention_settings

See ChannelRetentionSettings

channel_summary

See ChannelSummary

conversation_retention_settings

See ConversationRetentionSettings

create_attendee_error

See CreateAttendeeError

create_attendee_request_item

See CreateAttendeeRequestItem

credential

See Credential

dnis_emergency_calling_configuration

See DnisEmergencyCallingConfiguration

emergency_calling_configuration

See EmergencyCallingConfiguration

engine_transcribe_medical_settings

See EngineTranscribeMedicalSettings

engine_transcribe_settings

See EngineTranscribeSettings

events_configuration

See EventsConfiguration

geo_match_params

See GeoMatchParams

identity

See Identity

invite

See Invite

logging_configuration

See LoggingConfiguration

media_capture_pipeline

See MediaCapturePipeline

media_placement

See MediaPlacement

meeting

See Meeting

meeting_notification_configuration

See MeetingNotificationConfiguration

member

See Member

member_error

See MemberError

membership_item

See MembershipItem

messaging_session_endpoint

See MessagingSessionEndpoint

ordered_phone_number

See OrderedPhoneNumber

origination

See Origination

origination_route

See OriginationRoute

participant

See Participant

phone_number

See PhoneNumber

phone_number_association

See PhoneNumberAssociation

phone_number_capabilities

See PhoneNumberCapabilities

phone_number_country

See PhoneNumberCountry

phone_number_error

See PhoneNumberError

phone_number_order

See PhoneNumberOrder

proxy

See Proxy

proxy_session

See ProxySession

retention_settings

See RetentionSettings

room

See Room

room_membership

See RoomMembership

room_retention_settings

See RoomRetentionSettings

signin_delegate_group

See SigninDelegateGroup

sip_media_application

See SipMediaApplication

sip_media_application_call

See SipMediaApplicationCall

sip_media_application_endpoint

See SipMediaApplicationEndpoint

sip_media_application_logging_configuration

See SipMediaApplicationLoggingConfiguration

sip_rule

See SipRule

sip_rule_target_application

See SipRuleTargetApplication

streaming_configuration

See StreamingConfiguration

streaming_notification_target

See StreamingNotificationTarget

tag

See Tag

telephony_settings

See TelephonySettings

termination

See Termination

termination_health

See TerminationHealth

transcription_configuration

See TranscriptionConfiguration

update_phone_number_request_item

See UpdatePhoneNumberRequestItem

update_user_request_item

See UpdateUserRequestItem

user

See User

user_error

See UserError

user_settings

See UserSettings

voice_connector

See VoiceConnector

voice_connector_group

See VoiceConnectorGroup

voice_connector_item

See VoiceConnectorItem

voice_connector_settings

See VoiceConnectorSettings

Structs

Account

The Amazon Chime account details. An AWS account can have multiple Amazon Chime accounts.

AccountSettings

Settings related to the Amazon Chime account. This includes settings that start or stop

AlexaForBusinessMetadata

The Alexa for Business metadata associated with an Amazon Chime user, used to integrate Alexa for Business with a device.

AppInstance

The details of an AppInstance, an instance of an Amazon Chime SDK messaging application.

AppInstanceAdmin

The details of an AppInstanceAdmin.

AppInstanceAdminSummary

Summary of the details of an AppInstanceAdmin.

AppInstanceRetentionSettings

The details of the data-retention settings for an AppInstance.

AppInstanceStreamingConfiguration

The details of the streaming configuration of an AppInstance.

AppInstanceSummary

Summary of the data for an AppInstance.

AppInstanceUser

The details of an AppInstanceUser.

AppInstanceUserMembershipSummary

Summary of the membership details of an AppInstanceUser.

AppInstanceUserSummary

Summary of the details of an AppInstanceUser.

Attendee

BatchChannelMemberships

The membership information, including member ARNs, the channel ARN, and membership

BatchCreateChannelMembershipError

A list of failed member ARNs, error codes, and error messages.

Bot

A resource that allows Enterprise account administrators to configure an interface to receive events from Amazon Chime.

BusinessCallingSettings

The Amazon Chime Business Calling settings for the administrator's AWS account. Includes any

Channel

The details of a channel.

ChannelBan

The details of a channel ban.

ChannelBanSummary

Summary of the details of a ChannelBan.

ChannelMembership

The details of a channel member.

ChannelMembershipForAppInstanceUserSummary

Summary of the channel membership details of an AppInstanceUser.

ChannelMembershipSummary

Summary of the details of a ChannelMembership.

ChannelMessage

The details of a message in a channel.

ChannelMessageSummary

Summary of the messages in a Channel.

ChannelModeratedByAppInstanceUserSummary

Summary of the details of a moderated channel.

ChannelModerator

The details of a channel moderator.

ChannelModeratorSummary

Summary of the details of a ChannelModerator.

ChannelRetentionSettings

The details of the retention settings for a channel.

ChannelSummary

Summary of the details of a Channel.

ConversationRetentionSettings

The retention settings that determine how long to retain conversation messages for an Amazon Chime Enterprise account.

CreateAttendeeError

The list of errors returned when errors are encountered during the BatchCreateAttendee and CreateAttendee actions. This includes external user IDs, error codes, and error messages.

CreateAttendeeRequestItem

The Amazon Chime SDK attendee fields to create, used with the BatchCreateAttendee action.

Credential

The SIP credentials used to authenticate requests to your Amazon Chime Voice Connector.

DnisEmergencyCallingConfiguration

The Dialed Number Identification Service (DNIS) emergency calling configuration details associated with an Amazon Chime Voice Connector's emergency calling configuration.

EmergencyCallingConfiguration

The emergency calling configuration details associated with an Amazon Chime Voice Connector.

EngineTranscribeMedicalSettings

Settings specific to the Amazon Transcribe Medical engine.

EngineTranscribeSettings

Settings specific to the Amazon Transcribe engine.

EventsConfiguration

The configuration that allows a bot to receive outgoing events. Can be either an HTTPS endpoint or a Lambda function ARN.

GeoMatchParams

The country and area code for a proxy phone number in a proxy phone session.

Identity

The details of a user.

Invite

Invitation object returned after emailing users to invite them to join the Amazon Chime

LoggingConfiguration

The logging configuration associated with an Amazon Chime Voice Connector. Specifies whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

MediaCapturePipeline

A media capture pipeline object. A string consisting of an ID, source type, a source ARN, a sink type, and a sink ARN.

MediaPlacement

A set of endpoints used by clients to connect to the media service group for an Amazon Chime SDK meeting.

Meeting

A meeting created using the Amazon Chime SDK.

MeetingNotificationConfiguration

The resource target configurations for receiving Amazon Chime SDK meeting and attendee event

Member

The member details, such as email address, name, member ID, and member type.

MemberError

The list of errors returned when a member action results in an error.

MembershipItem

Membership details, such as member ID and member role.

MessagingSessionEndpoint

The websocket endpoint used to connect to Amazon Chime SDK messaging.

OrderedPhoneNumber

A phone number for which an order has been placed.

Origination

Origination settings enable your SIP hosts to receive inbound calls using your Amazon Chime

OriginationRoute

Origination routes define call distribution properties for your SIP hosts to receive inbound

Participant

The phone number and proxy phone number for a participant in an Amazon Chime Voice Connector proxy session.

PhoneNumber

A phone number used for Amazon Chime Business Calling or an Amazon Chime Voice

PhoneNumberAssociation

The phone number associations, such as Amazon Chime account ID, Amazon Chime user ID, Amazon

PhoneNumberCapabilities

The phone number capabilities for Amazon Chime Business Calling phone numbers, such as enabled

PhoneNumberCountry

The phone number country.

PhoneNumberError

If the phone number action fails for one or more of the phone numbers in the request, a list of

PhoneNumberOrder

The details of a phone number order created for Amazon Chime.

Proxy

The proxy configuration for an Amazon Chime Voice Connector.

ProxySession

The proxy session for an Amazon Chime Voice Connector.

RetentionSettings

The retention settings for an Amazon Chime Enterprise account that determine how long to retain items such as chat-room messages and chat-conversation messages.

Room

The Amazon Chime chat room details.

RoomMembership

The room membership details.

RoomRetentionSettings

The retention settings that determine how long to retain chat-room messages for an Amazon Chime Enterprise account.

SigninDelegateGroup

An Active Directory (AD) group whose members are granted permission to act as delegates.

SipMediaApplication

The details of the SIP media application, including name and endpoints. An AWS account can have multiple SIP media applications.

SipMediaApplicationCall

A Call instance for a SIP media application.

SipMediaApplicationEndpoint

The endpoint assigned to the SIP media application.

SipMediaApplicationLoggingConfiguration

Logging configuration of the SIP media application.

SipRule

The SIP rule details, including name, triggers, and target applications. An AWS account can have multiple SIP rules.

SipRuleTargetApplication

Target SIP media application and other details, such as priority and AWS Region, to be

StreamingConfiguration

The streaming configuration associated with an Amazon Chime Voice Connector. Specifies whether

StreamingNotificationTarget

The targeted recipient for a streaming configuration notification.

Tag

Describes a tag applied to a resource.

TelephonySettings

Settings that allow management of telephony permissions for an Amazon Chime user, such as

Termination

Termination settings enable your SIP hosts to make outbound calls using your Amazon Chime

TerminationHealth

The termination health details, including the source IP address and timestamp of the last

TranscriptionConfiguration

The configuration for the current transcription operation. Must contain EngineTranscribeSettings or EngineTranscribeMedicalSettings.

UpdatePhoneNumberRequestItem

The phone number ID, product type, or calling name fields to update, used with the

UpdateUserRequestItem

The user ID and user fields to update, used with the

User

The user on the Amazon Chime account.

UserError

The list of errors returned when errors are encountered during the BatchSuspendUser, BatchUnsuspendUser, or

UserSettings

Settings associated with an Amazon Chime user, including inbound and outbound calling and text

VoiceConnector

The Amazon Chime Voice Connector configuration, including outbound host name and encryption

VoiceConnectorGroup

The Amazon Chime Voice Connector group configuration, including associated Amazon Chime Voice

VoiceConnectorItem

For Amazon Chime Voice Connector groups, the Amazon Chime Voice Connectors to which to route inbound calls. Includes priority configuration settings. Limit: 3

VoiceConnectorSettings

The Amazon Chime Voice Connector settings. Includes any Amazon S3 buckets designated for

Enums

AccountStatus

AccountType

AppInstanceDataType

BotType

CallingNameStatus

Capability

ChannelMembershipType

ChannelMessagePersistenceType

ChannelMessageType

ChannelMode

ChannelPrivacy

EmailStatus

ErrorCode

GeoMatchLevel

InviteStatus

License

MediaPipelineSinkType

MediaPipelineSourceType

MediaPipelineStatus

MemberType

NotificationTarget

NumberSelectionBehavior

OrderedPhoneNumberStatus

OriginationRouteProtocol

PhoneNumberAssociationName

PhoneNumberOrderStatus

PhoneNumberProductType

PhoneNumberStatus

PhoneNumberType

ProxySessionStatus

RegistrationStatus

RoomMembershipRole

SipRuleTriggerType

SortOrder

TranscribeLanguageCode

TranscribeMedicalLanguageCode

TranscribeMedicalRegion

TranscribeMedicalSpecialty

TranscribeMedicalType

TranscribeRegion

TranscribeVocabularyFilterMethod

UserType

VoiceConnectorAwsRegion