Module aws_sdk_chime::operation

Structs

AssociatePhoneNumberWithUser

Associates a phone number with the specified Amazon Chime user.

AssociatePhoneNumbersWithVoiceConnector

Associates phone numbers with the specified Amazon Chime Voice Connector.

AssociatePhoneNumbersWithVoiceConnectorGroup

Associates phone numbers with the specified Amazon Chime Voice Connector group.

AssociateSigninDelegateGroupsWithAccount

Associates the specified sign-in delegate groups with the specified Amazon Chime account.

BatchCreateAttendee

BatchCreateChannelMembership

Adds a specified number of users to a channel.

BatchCreateRoomMembership

Adds up to 50 members to a chat room in an Amazon Chime Enterprise account. Members can be users or bots. The member role designates whether the member is a chat room administrator or a general chat room member.

BatchDeletePhoneNumber

BatchSuspendUser

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime

BatchUnsuspendUser

Removes the suspension from up to 50 previously suspended users for the specified Amazon

BatchUpdatePhoneNumber

Updates phone number product types or calling names. You can update one attribute at a time for each UpdatePhoneNumberRequestItem. For example, you can update the product type or the calling name.

BatchUpdateUser

Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account. Currently, only LicenseType updates are supported for this action.

CreateAccount

Creates an Amazon Chime account under the administrator's AWS account. Only Team

CreateAppInstance

Creates an Amazon Chime SDK messaging AppInstance under an AWS account. Only SDK messaging customers use this API.

CreateAppInstanceAdmin

Promotes an AppInstanceUser to an AppInstanceAdmin. The promoted user can perform the following actions.

CreateAppInstanceUser

Creates a user under an Amazon Chime AppInstance. The request consists of a unique appInstanceUserId and

CreateAttendee

CreateBot

Creates a bot for an Amazon Chime Enterprise account.

CreateChannel

Creates a channel to which you can add users and send messages.

CreateChannelBan

Permanently bans a member from a channel. Moderators can't add banned members to a

CreateChannelMembership

Adds a user to a channel. The InvitedBy response field is derived from the

CreateChannelModerator

Creates a new ChannelModerator. A channel moderator can:

CreateMediaCapturePipeline

Creates a media capture pipeline.

CreateMeeting

CreateMeetingDialOut

Uses the join token and call metadata in a meeting request (From number, To number, and so forth) to initiate an outbound call to a public

CreateMeetingWithAttendees

CreatePhoneNumberOrder

Creates an order for phone numbers to be provisioned. For toll-free numbers, you cannot use the Amazon Chime Business Calling product type.

CreateProxySession

Creates a proxy session on the specified Amazon Chime Voice Connector for the specified participant phone numbers.

CreateRoom

Creates a chat room for the specified Amazon Chime Enterprise account.

CreateRoomMembership

Adds a member to a chat room in an Amazon Chime Enterprise account. A member can be either a user or a bot. The member role designates whether the member is a chat room administrator or a general chat room member.

CreateSipMediaApplication

Creates a SIP media application.

CreateSipMediaApplicationCall

Creates an outbound call to a phone number from the phone number specified in the request, and it invokes the endpoint of the specified

CreateSipRule

Creates a SIP rule which can be used to run a SIP media application as a target for a specific trigger type.

CreateUser

Creates a user under the specified Amazon Chime account.

CreateVoiceConnector

Creates an Amazon Chime Voice Connector under the administrator's AWS account. You can choose to create an Amazon Chime Voice Connector in a specific AWS Region.

CreateVoiceConnectorGroup

Creates an Amazon Chime Voice Connector group under the administrator's AWS account. You can

DeleteAccount

Deletes the specified Amazon Chime account. You must suspend all users before deleting

DeleteAppInstance

Deletes an AppInstance and all associated data asynchronously.

DeleteAppInstanceAdmin

Demotes an AppInstanceAdmin to an AppInstanceUser. This action does not delete the user.

DeleteAppInstanceStreamingConfigurations

Deletes the streaming configurations of an AppInstance.

DeleteAppInstanceUser

Deletes an AppInstanceUser.

DeleteAttendee

Deletes an attendee from the specified Amazon Chime SDK meeting and deletes their

DeleteChannel

Immediately makes a channel and its memberships inaccessible and marks them for

DeleteChannelBan

Removes a user from a channel's ban list.

DeleteChannelMembership

Removes a member from a channel.

DeleteChannelMessage

Deletes a channel message. Only admins can perform this action. Deletion makes messages

DeleteChannelModerator

Deletes a channel moderator.

DeleteEventsConfiguration

Deletes the events configuration that allows a bot to receive outgoing events.

DeleteMediaCapturePipeline

Deletes the media capture pipeline.

DeleteMeeting

Deletes the specified Amazon Chime SDK meeting. The operation deletes all attendees, disconnects all clients, and prevents new clients from

DeletePhoneNumber

Moves the specified phone number into the Deletion queue. A

DeleteProxySession

Deletes the specified proxy session from the specified Amazon Chime Voice Connector.

DeleteRoom

Deletes a chat room in an Amazon Chime Enterprise account.

DeleteRoomMembership

Removes a member from a chat room in an Amazon Chime Enterprise account.

DeleteSipMediaApplication

Deletes a SIP media application.

DeleteSipRule

Deletes a SIP rule. You must disable a SIP rule before you can delete it.

DeleteVoiceConnector

Deletes the specified Amazon Chime Voice Connector. Any phone numbers associated with the

DeleteVoiceConnectorEmergencyCallingConfiguration

Deletes the emergency calling configuration details from the specified Amazon Chime Voice Connector.

DeleteVoiceConnectorGroup

Deletes the specified Amazon Chime Voice Connector group. Any

DeleteVoiceConnectorOrigination

Deletes the origination settings for the specified Amazon Chime Voice Connector.

DeleteVoiceConnectorProxy

Deletes the proxy configuration from the specified Amazon Chime Voice Connector.

DeleteVoiceConnectorStreamingConfiguration

Deletes the streaming configuration for the specified Amazon Chime Voice Connector.

DeleteVoiceConnectorTermination

Deletes the termination settings for the specified Amazon Chime Voice Connector.

DeleteVoiceConnectorTerminationCredentials

Deletes the specified SIP credentials used by your equipment to authenticate during call termination.

DescribeAppInstance

Returns the full details of an AppInstance.

DescribeAppInstanceAdmin

Returns the full details of an AppInstanceAdmin.

DescribeAppInstanceUser

Returns the full details of an AppInstanceUser.

DescribeChannel

Returns the full details of a channel in an Amazon Chime

DescribeChannelBan

Returns the full details of a channel ban.

DescribeChannelMembership

Returns the full details of a user's channel membership.

DescribeChannelMembershipForAppInstanceUser

Returns the details of a channel based on the membership of the specified

DescribeChannelModeratedByAppInstanceUser

Returns the full details of a channel moderated by the specified

DescribeChannelModerator

Returns the full details of a single ChannelModerator.

DisassociatePhoneNumberFromUser

Disassociates the primary provisioned phone number from the specified Amazon Chime user.

DisassociatePhoneNumbersFromVoiceConnector

Disassociates the specified phone numbers from the specified Amazon Chime Voice Connector.

DisassociatePhoneNumbersFromVoiceConnectorGroup

Disassociates the specified phone numbers from the specified Amazon Chime Voice Connector group.

DisassociateSigninDelegateGroupsFromAccount

Disassociates the specified sign-in delegate groups from the specified Amazon Chime account.

GetAccount

Retrieves details for the specified Amazon Chime account, such as account type and supported

GetAccountSettings

Retrieves account settings for the specified Amazon Chime account ID, such as remote control

GetAppInstanceRetentionSettings

Gets the retention settings for an AppInstance.

GetAppInstanceStreamingConfigurations

Gets the streaming settings for an AppInstance.

GetAttendee

GetBot

Retrieves details for the specified bot, such as bot email address, bot type, status, and display name.

GetChannelMessage

Gets the full details of a channel message.

GetEventsConfiguration

Gets details for an events configuration that allows a bot to receive outgoing events, such as an HTTPS endpoint or Lambda function ARN.

GetGlobalSettings

Retrieves global settings for the administrator's AWS account, such as Amazon Chime Business

GetMediaCapturePipeline

Gets an existing media capture pipeline.

GetMeeting

GetMessagingSessionEndpoint

The details of the endpoint for the messaging session.

GetPhoneNumber

Retrieves details for the specified phone number ID, such as associations, capabilities, and product type.

GetPhoneNumberOrder

Retrieves details for the specified phone number order, such as the order creation timestamp, phone

GetPhoneNumberSettings

Retrieves the phone number settings for the administrator's AWS account, such as the default outbound calling name.

GetProxySession

Gets the specified proxy session details for the specified Amazon Chime Voice Connector.

GetRetentionSettings

GetRoom

Retrieves room details, such as the room name, for a room in an Amazon Chime Enterprise account.

GetSipMediaApplication

Retrieves the information for a SIP media application, including name, AWS Region, and endpoints.

GetSipMediaApplicationLoggingConfiguration

Returns the logging configuration for the specified SIP media application.

GetSipRule

Retrieves the details of a SIP rule, such as the rule ID, name, triggers, and target endpoints.

GetUser

Retrieves details for the specified user ID, such as primary email address, license type,and personal meeting PIN.

GetUserSettings

Retrieves settings for the specified user ID, such as any associated phone number settings.

GetVoiceConnector

Retrieves details for the specified Amazon Chime Voice Connector, such as timestamps,name, outbound host, and encryption requirements.

GetVoiceConnectorEmergencyCallingConfiguration

Gets the emergency calling configuration details for the specified Amazon Chime Voice Connector.

GetVoiceConnectorGroup

GetVoiceConnectorLoggingConfiguration

Retrieves the logging configuration details for the specified Amazon Chime Voice Connector. Shows whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

GetVoiceConnectorOrigination

Retrieves origination setting details for the specified Amazon Chime Voice Connector.

GetVoiceConnectorProxy

Gets the proxy configuration details for the specified Amazon Chime Voice Connector.

GetVoiceConnectorStreamingConfiguration

Retrieves the streaming configuration details for the specified Amazon Chime Voice Connector.

GetVoiceConnectorTermination

Retrieves termination setting details for the specified Amazon Chime Voice Connector.

GetVoiceConnectorTerminationHealth

Retrieves information about the last time a SIP OPTIONS ping was received from

InviteUsers

Sends email to a maximum of 50 users, inviting them to the specified Amazon Chime

ListAccounts

Lists the Amazon Chime accounts under the administrator's AWS account. You can filter accounts

ListAppInstanceAdmins

Returns a list of the administrators in the AppInstance.

ListAppInstanceUsers

List all AppInstanceUsers

ListAppInstances

Lists all Amazon Chime AppInstances created under a single AWS account.

ListAttendeeTags

Lists the tags applied to an Amazon Chime SDK attendee resource.

ListAttendees

ListBots

Lists the bots associated with the administrator's Amazon Chime Enterprise account ID.

ListChannelBans

Lists all the users banned from a particular channel.

ListChannelMemberships

Lists all channel memberships in a channel.

ListChannelMembershipsForAppInstanceUser

Lists all channels that a particular AppInstanceUser is a part of. Only an

ListChannelMessages

List all the messages in a channel. Returns a paginated list of

ListChannelModerators

Lists all the moderators for a channel.

ListChannels

Lists all Channels created under a single Chime App as a paginated list. You can specify

ListChannelsModeratedByAppInstanceUser

A list of the channels moderated by an AppInstanceUser.

ListMediaCapturePipelines

Returns a list of media capture pipelines.

ListMeetingTags

Lists the tags applied to an Amazon Chime SDK meeting resource.

ListMeetings

ListPhoneNumberOrders

Lists the phone number orders for the administrator's Amazon Chime account.

ListPhoneNumbers

Lists the phone numbers for the specified Amazon Chime account, Amazon Chime user, Amazon Chime Voice Connector, or Amazon Chime Voice Connector group.

ListProxySessions

Lists the proxy sessions for the specified Amazon Chime Voice Connector.

ListRoomMemberships

Lists the membership details for the specified room in an Amazon Chime Enterprise account,

ListRooms

Lists the room details for the specified Amazon Chime Enterprise account. Optionally, filter the results by a member ID (user ID or bot ID) to see a list of rooms that the member belongs to.

ListSipMediaApplications

Lists the SIP media applications under the administrator's AWS account.

ListSipRules

Lists the SIP rules under the administrator's AWS account.

ListSupportedPhoneNumberCountries

Lists supported phone number countries.

ListTagsForResource

Lists the tags applied to an Amazon Chime SDK meeting resource.

ListUsers

Lists the users that belong to the specified Amazon Chime account. You can specify an email

ListVoiceConnectorGroups

Lists the Amazon Chime Voice Connector groups for the administrator's AWS account.

ListVoiceConnectorTerminationCredentials

Lists the SIP credentials for the specified Amazon Chime Voice Connector.

ListVoiceConnectors

Lists the Amazon Chime Voice Connectors for the administrator's AWS account.

LogoutUser

Logs out the specified user from all of the devices they are currently logged into.

PutAppInstanceRetentionSettings

Sets the amount of time in days that a given AppInstance retains data.

PutAppInstanceStreamingConfigurations

The data streaming configurations of an AppInstance.

PutEventsConfiguration

Creates an events configuration that allows a bot to receive outgoing events sent by Amazon

PutRetentionSettings

PutSipMediaApplicationLoggingConfiguration

Updates the logging configuration for the specified SIP media application.

PutVoiceConnectorEmergencyCallingConfiguration

Puts emergency calling configuration details to the specified Amazon Chime Voice Connector, such as emergency phone numbers and calling countries. Origination and termination settings must be enabled for

PutVoiceConnectorLoggingConfiguration

Adds a logging configuration for the specified Amazon Chime Voice Connector. The logging configuration specifies whether SIP message logs are enabled for sending to Amazon CloudWatch Logs.

PutVoiceConnectorOrigination

Adds origination settings for the specified Amazon Chime Voice Connector.

PutVoiceConnectorProxy

Puts the specified proxy configuration to the specified Amazon Chime Voice Connector.

PutVoiceConnectorStreamingConfiguration

Adds a streaming configuration for the specified Amazon Chime Voice Connector. The streaming

PutVoiceConnectorTermination

Adds termination settings for the specified Amazon Chime Voice Connector.

PutVoiceConnectorTerminationCredentials

Adds termination SIP credentials for the specified Amazon Chime Voice Connector.

RedactChannelMessage

Redacts message content, but not metadata. The message exists in the back end, but the

RedactConversationMessage

Redacts the specified message from the specified Amazon Chime conversation.

RedactRoomMessage

Redacts the specified message from the specified Amazon Chime channel.

RegenerateSecurityToken

Regenerates the security token for a bot.

ResetPersonalPIN

Resets the personal meeting PIN for the specified user on an Amazon Chime account. Returns

RestorePhoneNumber

Moves a phone number from the Deletion queue back into the

SearchAvailablePhoneNumbers

Searches for phone numbers that can be ordered. For US numbers, provide at least one of

SendChannelMessage

Sends a message to a particular channel that the member is a part of.

StartMeetingTranscription

Start transcription for the specified meetingId.

StopMeetingTranscription

Stops transcription for the specified meetingId.

TagAttendee

Applies the specified tags to the specified Amazon Chime SDK attendee.

TagMeeting

Applies the specified tags to the specified Amazon Chime SDK meeting.

TagResource

Applies the specified tags to the specified Amazon Chime SDK meeting resource.

UntagAttendee

Untags the specified tags from the specified Amazon Chime SDK attendee.

UntagMeeting

Untags the specified tags from the specified Amazon Chime SDK meeting.

UntagResource

Untags the specified tags from the specified Amazon Chime SDK meeting resource.

UpdateAccount

Updates account details for the specified Amazon Chime account. Currently, only account name and default license updates are supported for this action.

UpdateAccountSettings

Updates the settings for the specified Amazon Chime account. You can update settings for

UpdateAppInstance

Updates AppInstance metadata.

UpdateAppInstanceUser

Updates the details of an AppInstanceUser. You can update names and metadata.

UpdateBot

Updates the status of the specified bot, such as starting or stopping the bot from running in your Amazon Chime Enterprise account.

UpdateChannel

Update a channel's attributes.

UpdateChannelMessage

Updates the content of a message.

UpdateChannelReadMarker

The details of the time when a user last read messages in a channel.

UpdateGlobalSettings

Updates global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

UpdatePhoneNumber

Updates phone number details, such as product type or calling name, for the specified phone number ID. You can update one phone number detail at a time. For example, you can update either the product type or the calling name in one action.

UpdatePhoneNumberSettings

Updates the phone number settings for the administrator's AWS account, such as the default

UpdateProxySession

Updates the specified proxy session details, such as voice or SMS capabilities.

UpdateRoom

Updates room details, such as the room name, for a room in an Amazon Chime Enterprise account.

UpdateRoomMembership

Updates room membership details, such as the member role, for a room in an Amazon Chime

UpdateSipMediaApplication

Updates the details of the specified SIP media application.

UpdateSipMediaApplicationCall

Allows you to trigger a Lambda function at any time while a call is active, and replace the current actions with new actions returned by the invocation.

UpdateSipRule

Updates the details of the specified SIP rule.

UpdateUser

Updates user details for a specified user ID. Currently, only LicenseType updates are supported for this action.

UpdateUserSettings

Updates the settings for the specified user, such as phone number settings.

UpdateVoiceConnector

Updates details for the specified Amazon Chime Voice Connector.

UpdateVoiceConnectorGroup

Updates details of the specified Amazon Chime Voice Connector group, such as the name and