Struct aws_sdk_chime::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for UCBuzzConsoleService.

This client allows ergonomic access to a UCBuzzConsoleService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_phone_numbers_with_voice_connector(
&self
) -> AssociatePhoneNumbersWithVoiceConnector<C, M, R>

Constructs a fluent builder for the AssociatePhoneNumbersWithVoiceConnector operation.

See AssociatePhoneNumbersWithVoiceConnector for more information about the operation and its arguments.

pub fn associate_phone_numbers_with_voice_connector_group(
&self
) -> AssociatePhoneNumbersWithVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the AssociatePhoneNumbersWithVoiceConnectorGroup operation.

See AssociatePhoneNumbersWithVoiceConnectorGroup for more information about the operation and its arguments.

pub fn associate_phone_number_with_user(
&self
) -> AssociatePhoneNumberWithUser<C, M, R>

Constructs a fluent builder for the AssociatePhoneNumberWithUser operation.

See AssociatePhoneNumberWithUser for more information about the operation and its arguments.

pub fn associate_signin_delegate_groups_with_account(
&self
) -> AssociateSigninDelegateGroupsWithAccount<C, M, R>

Constructs a fluent builder for the AssociateSigninDelegateGroupsWithAccount operation.

See AssociateSigninDelegateGroupsWithAccount for more information about the operation and its arguments.

pub fn batch_create_attendee(&self) -> BatchCreateAttendee<C, M, R>

Constructs a fluent builder for the BatchCreateAttendee operation.

See BatchCreateAttendee for more information about the operation and its arguments.

pub fn batch_create_channel_membership(
&self
) -> BatchCreateChannelMembership<C, M, R>

Constructs a fluent builder for the BatchCreateChannelMembership operation.

See BatchCreateChannelMembership for more information about the operation and its arguments.

pub fn batch_create_room_membership(&self) -> BatchCreateRoomMembership<C, M, R>

Constructs a fluent builder for the BatchCreateRoomMembership operation.

See BatchCreateRoomMembership for more information about the operation and its arguments.

pub fn batch_delete_phone_number(&self) -> BatchDeletePhoneNumber<C, M, R>

Constructs a fluent builder for the BatchDeletePhoneNumber operation.

See BatchDeletePhoneNumber for more information about the operation and its arguments.

pub fn batch_suspend_user(&self) -> BatchSuspendUser<C, M, R>

Constructs a fluent builder for the BatchSuspendUser operation.

See BatchSuspendUser for more information about the operation and its arguments.

pub fn batch_unsuspend_user(&self) -> BatchUnsuspendUser<C, M, R>

Constructs a fluent builder for the BatchUnsuspendUser operation.

See BatchUnsuspendUser for more information about the operation and its arguments.

pub fn batch_update_phone_number(&self) -> BatchUpdatePhoneNumber<C, M, R>

Constructs a fluent builder for the BatchUpdatePhoneNumber operation.

See BatchUpdatePhoneNumber for more information about the operation and its arguments.

pub fn batch_update_user(&self) -> BatchUpdateUser<C, M, R>

Constructs a fluent builder for the BatchUpdateUser operation.

See BatchUpdateUser for more information about the operation and its arguments.

pub fn create_account(&self) -> CreateAccount<C, M, R>

Constructs a fluent builder for the CreateAccount operation.

See CreateAccount for more information about the operation and its arguments.

pub fn create_app_instance(&self) -> CreateAppInstance<C, M, R>

Constructs a fluent builder for the CreateAppInstance operation.

See CreateAppInstance for more information about the operation and its arguments.

pub fn create_app_instance_admin(&self) -> CreateAppInstanceAdmin<C, M, R>

Constructs a fluent builder for the CreateAppInstanceAdmin operation.

See CreateAppInstanceAdmin for more information about the operation and its arguments.

pub fn create_app_instance_user(&self) -> CreateAppInstanceUser<C, M, R>

Constructs a fluent builder for the CreateAppInstanceUser operation.

See CreateAppInstanceUser for more information about the operation and its arguments.

pub fn create_attendee(&self) -> CreateAttendee<C, M, R>

Constructs a fluent builder for the CreateAttendee operation.

See CreateAttendee for more information about the operation and its arguments.

pub fn create_bot(&self) -> CreateBot<C, M, R>

Constructs a fluent builder for the CreateBot operation.

See CreateBot for more information about the operation and its arguments.

pub fn create_channel(&self) -> CreateChannel<C, M, R>

Constructs a fluent builder for the CreateChannel operation.

See CreateChannel for more information about the operation and its arguments.

pub fn create_channel_ban(&self) -> CreateChannelBan<C, M, R>

Constructs a fluent builder for the CreateChannelBan operation.

See CreateChannelBan for more information about the operation and its arguments.

pub fn create_channel_membership(&self) -> CreateChannelMembership<C, M, R>

Constructs a fluent builder for the CreateChannelMembership operation.

See CreateChannelMembership for more information about the operation and its arguments.

pub fn create_channel_moderator(&self) -> CreateChannelModerator<C, M, R>

Constructs a fluent builder for the CreateChannelModerator operation.

See CreateChannelModerator for more information about the operation and its arguments.

pub fn create_media_capture_pipeline(
&self
) -> CreateMediaCapturePipeline<C, M, R>

Constructs a fluent builder for the CreateMediaCapturePipeline operation.

See CreateMediaCapturePipeline for more information about the operation and its arguments.

pub fn create_meeting(&self) -> CreateMeeting<C, M, R>

Constructs a fluent builder for the CreateMeeting operation.

See CreateMeeting for more information about the operation and its arguments.

pub fn create_meeting_dial_out(&self) -> CreateMeetingDialOut<C, M, R>

Constructs a fluent builder for the CreateMeetingDialOut operation.

See CreateMeetingDialOut for more information about the operation and its arguments.

pub fn create_meeting_with_attendees(
&self
) -> CreateMeetingWithAttendees<C, M, R>

Constructs a fluent builder for the CreateMeetingWithAttendees operation.

See CreateMeetingWithAttendees for more information about the operation and its arguments.

pub fn create_phone_number_order(&self) -> CreatePhoneNumberOrder<C, M, R>

Constructs a fluent builder for the CreatePhoneNumberOrder operation.

See CreatePhoneNumberOrder for more information about the operation and its arguments.

pub fn create_proxy_session(&self) -> CreateProxySession<C, M, R>

Constructs a fluent builder for the CreateProxySession operation.

See CreateProxySession for more information about the operation and its arguments.

pub fn create_room(&self) -> CreateRoom<C, M, R>

Constructs a fluent builder for the CreateRoom operation.

See CreateRoom for more information about the operation and its arguments.

pub fn create_room_membership(&self) -> CreateRoomMembership<C, M, R>

Constructs a fluent builder for the CreateRoomMembership operation.

See CreateRoomMembership for more information about the operation and its arguments.

pub fn create_sip_media_application(&self) -> CreateSipMediaApplication<C, M, R>

Constructs a fluent builder for the CreateSipMediaApplication operation.

See CreateSipMediaApplication for more information about the operation and its arguments.

pub fn create_sip_media_application_call(
&self
) -> CreateSipMediaApplicationCall<C, M, R>

Constructs a fluent builder for the CreateSipMediaApplicationCall operation.

See CreateSipMediaApplicationCall for more information about the operation and its arguments.

pub fn create_sip_rule(&self) -> CreateSipRule<C, M, R>

Constructs a fluent builder for the CreateSipRule operation.

See CreateSipRule for more information about the operation and its arguments.

pub fn create_user(&self) -> CreateUser<C, M, R>

Constructs a fluent builder for the CreateUser operation.

See CreateUser for more information about the operation and its arguments.

pub fn create_voice_connector(&self) -> CreateVoiceConnector<C, M, R>

Constructs a fluent builder for the CreateVoiceConnector operation.

See CreateVoiceConnector for more information about the operation and its arguments.

pub fn create_voice_connector_group(&self) -> CreateVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the CreateVoiceConnectorGroup operation.

See CreateVoiceConnectorGroup for more information about the operation and its arguments.

pub fn delete_account(&self) -> DeleteAccount<C, M, R>

Constructs a fluent builder for the DeleteAccount operation.

See DeleteAccount for more information about the operation and its arguments.

pub fn delete_app_instance(&self) -> DeleteAppInstance<C, M, R>

Constructs a fluent builder for the DeleteAppInstance operation.

See DeleteAppInstance for more information about the operation and its arguments.

pub fn delete_app_instance_admin(&self) -> DeleteAppInstanceAdmin<C, M, R>

Constructs a fluent builder for the DeleteAppInstanceAdmin operation.

See DeleteAppInstanceAdmin for more information about the operation and its arguments.

pub fn delete_app_instance_streaming_configurations(
&self
) -> DeleteAppInstanceStreamingConfigurations<C, M, R>

Constructs a fluent builder for the DeleteAppInstanceStreamingConfigurations operation.

See DeleteAppInstanceStreamingConfigurations for more information about the operation and its arguments.

pub fn delete_app_instance_user(&self) -> DeleteAppInstanceUser<C, M, R>

Constructs a fluent builder for the DeleteAppInstanceUser operation.

See DeleteAppInstanceUser for more information about the operation and its arguments.

pub fn delete_attendee(&self) -> DeleteAttendee<C, M, R>

Constructs a fluent builder for the DeleteAttendee operation.

See DeleteAttendee for more information about the operation and its arguments.

pub fn delete_channel(&self) -> DeleteChannel<C, M, R>

Constructs a fluent builder for the DeleteChannel operation.

See DeleteChannel for more information about the operation and its arguments.

pub fn delete_channel_ban(&self) -> DeleteChannelBan<C, M, R>

Constructs a fluent builder for the DeleteChannelBan operation.

See DeleteChannelBan for more information about the operation and its arguments.

pub fn delete_channel_membership(&self) -> DeleteChannelMembership<C, M, R>

Constructs a fluent builder for the DeleteChannelMembership operation.

See DeleteChannelMembership for more information about the operation and its arguments.

pub fn delete_channel_message(&self) -> DeleteChannelMessage<C, M, R>

Constructs a fluent builder for the DeleteChannelMessage operation.

See DeleteChannelMessage for more information about the operation and its arguments.

pub fn delete_channel_moderator(&self) -> DeleteChannelModerator<C, M, R>

Constructs a fluent builder for the DeleteChannelModerator operation.

See DeleteChannelModerator for more information about the operation and its arguments.

pub fn delete_events_configuration(&self) -> DeleteEventsConfiguration<C, M, R>

Constructs a fluent builder for the DeleteEventsConfiguration operation.

See DeleteEventsConfiguration for more information about the operation and its arguments.

pub fn delete_media_capture_pipeline(
&self
) -> DeleteMediaCapturePipeline<C, M, R>

Constructs a fluent builder for the DeleteMediaCapturePipeline operation.

See DeleteMediaCapturePipeline for more information about the operation and its arguments.

pub fn delete_meeting(&self) -> DeleteMeeting<C, M, R>

Constructs a fluent builder for the DeleteMeeting operation.

See DeleteMeeting for more information about the operation and its arguments.

pub fn delete_phone_number(&self) -> DeletePhoneNumber<C, M, R>

Constructs a fluent builder for the DeletePhoneNumber operation.

See DeletePhoneNumber for more information about the operation and its arguments.

pub fn delete_proxy_session(&self) -> DeleteProxySession<C, M, R>

Constructs a fluent builder for the DeleteProxySession operation.

See DeleteProxySession for more information about the operation and its arguments.

pub fn delete_room(&self) -> DeleteRoom<C, M, R>

Constructs a fluent builder for the DeleteRoom operation.

See DeleteRoom for more information about the operation and its arguments.

pub fn delete_room_membership(&self) -> DeleteRoomMembership<C, M, R>

Constructs a fluent builder for the DeleteRoomMembership operation.

See DeleteRoomMembership for more information about the operation and its arguments.

pub fn delete_sip_media_application(&self) -> DeleteSipMediaApplication<C, M, R>

Constructs a fluent builder for the DeleteSipMediaApplication operation.

See DeleteSipMediaApplication for more information about the operation and its arguments.

pub fn delete_sip_rule(&self) -> DeleteSipRule<C, M, R>

Constructs a fluent builder for the DeleteSipRule operation.

See DeleteSipRule for more information about the operation and its arguments.

pub fn delete_voice_connector(&self) -> DeleteVoiceConnector<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnector operation.

See DeleteVoiceConnector for more information about the operation and its arguments.

pub fn delete_voice_connector_emergency_calling_configuration(
&self
) -> DeleteVoiceConnectorEmergencyCallingConfiguration<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorEmergencyCallingConfiguration operation.

See DeleteVoiceConnectorEmergencyCallingConfiguration for more information about the operation and its arguments.

pub fn delete_voice_connector_group(&self) -> DeleteVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorGroup operation.

See DeleteVoiceConnectorGroup for more information about the operation and its arguments.

pub fn delete_voice_connector_origination(
&self
) -> DeleteVoiceConnectorOrigination<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorOrigination operation.

See DeleteVoiceConnectorOrigination for more information about the operation and its arguments.

pub fn delete_voice_connector_proxy(&self) -> DeleteVoiceConnectorProxy<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorProxy operation.

See DeleteVoiceConnectorProxy for more information about the operation and its arguments.

pub fn delete_voice_connector_streaming_configuration(
&self
) -> DeleteVoiceConnectorStreamingConfiguration<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorStreamingConfiguration operation.

See DeleteVoiceConnectorStreamingConfiguration for more information about the operation and its arguments.

pub fn delete_voice_connector_termination(
&self
) -> DeleteVoiceConnectorTermination<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorTermination operation.

See DeleteVoiceConnectorTermination for more information about the operation and its arguments.

pub fn delete_voice_connector_termination_credentials(
&self
) -> DeleteVoiceConnectorTerminationCredentials<C, M, R>

Constructs a fluent builder for the DeleteVoiceConnectorTerminationCredentials operation.

See DeleteVoiceConnectorTerminationCredentials for more information about the operation and its arguments.

pub fn describe_app_instance(&self) -> DescribeAppInstance<C, M, R>

Constructs a fluent builder for the DescribeAppInstance operation.

See DescribeAppInstance for more information about the operation and its arguments.

pub fn describe_app_instance_admin(&self) -> DescribeAppInstanceAdmin<C, M, R>

Constructs a fluent builder for the DescribeAppInstanceAdmin operation.

See DescribeAppInstanceAdmin for more information about the operation and its arguments.

pub fn describe_app_instance_user(&self) -> DescribeAppInstanceUser<C, M, R>

Constructs a fluent builder for the DescribeAppInstanceUser operation.

See DescribeAppInstanceUser for more information about the operation and its arguments.

pub fn describe_channel(&self) -> DescribeChannel<C, M, R>

Constructs a fluent builder for the DescribeChannel operation.

See DescribeChannel for more information about the operation and its arguments.

pub fn describe_channel_ban(&self) -> DescribeChannelBan<C, M, R>

Constructs a fluent builder for the DescribeChannelBan operation.

See DescribeChannelBan for more information about the operation and its arguments.

pub fn describe_channel_membership(&self) -> DescribeChannelMembership<C, M, R>

Constructs a fluent builder for the DescribeChannelMembership operation.

See DescribeChannelMembership for more information about the operation and its arguments.

pub fn describe_channel_membership_for_app_instance_user(
&self
) -> DescribeChannelMembershipForAppInstanceUser<C, M, R>

Constructs a fluent builder for the DescribeChannelMembershipForAppInstanceUser operation.

See DescribeChannelMembershipForAppInstanceUser for more information about the operation and its arguments.

pub fn describe_channel_moderated_by_app_instance_user(
&self
) -> DescribeChannelModeratedByAppInstanceUser<C, M, R>

Constructs a fluent builder for the DescribeChannelModeratedByAppInstanceUser operation.

See DescribeChannelModeratedByAppInstanceUser for more information about the operation and its arguments.

pub fn describe_channel_moderator(&self) -> DescribeChannelModerator<C, M, R>

Constructs a fluent builder for the DescribeChannelModerator operation.

See DescribeChannelModerator for more information about the operation and its arguments.

pub fn disassociate_phone_number_from_user(
&self
) -> DisassociatePhoneNumberFromUser<C, M, R>

Constructs a fluent builder for the DisassociatePhoneNumberFromUser operation.

See DisassociatePhoneNumberFromUser for more information about the operation and its arguments.

pub fn disassociate_phone_numbers_from_voice_connector(
&self
) -> DisassociatePhoneNumbersFromVoiceConnector<C, M, R>

Constructs a fluent builder for the DisassociatePhoneNumbersFromVoiceConnector operation.

See DisassociatePhoneNumbersFromVoiceConnector for more information about the operation and its arguments.

pub fn disassociate_phone_numbers_from_voice_connector_group(
&self
) -> DisassociatePhoneNumbersFromVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the DisassociatePhoneNumbersFromVoiceConnectorGroup operation.

See DisassociatePhoneNumbersFromVoiceConnectorGroup for more information about the operation and its arguments.

pub fn disassociate_signin_delegate_groups_from_account(
&self
) -> DisassociateSigninDelegateGroupsFromAccount<C, M, R>

Constructs a fluent builder for the DisassociateSigninDelegateGroupsFromAccount operation.

See DisassociateSigninDelegateGroupsFromAccount for more information about the operation and its arguments.

pub fn get_account(&self) -> GetAccount<C, M, R>

Constructs a fluent builder for the GetAccount operation.

See GetAccount for more information about the operation and its arguments.

pub fn get_account_settings(&self) -> GetAccountSettings<C, M, R>

Constructs a fluent builder for the GetAccountSettings operation.

See GetAccountSettings for more information about the operation and its arguments.

pub fn get_app_instance_retention_settings(
&self
) -> GetAppInstanceRetentionSettings<C, M, R>

Constructs a fluent builder for the GetAppInstanceRetentionSettings operation.

See GetAppInstanceRetentionSettings for more information about the operation and its arguments.

pub fn get_app_instance_streaming_configurations(
&self
) -> GetAppInstanceStreamingConfigurations<C, M, R>

Constructs a fluent builder for the GetAppInstanceStreamingConfigurations operation.

See GetAppInstanceStreamingConfigurations for more information about the operation and its arguments.

pub fn get_attendee(&self) -> GetAttendee<C, M, R>

Constructs a fluent builder for the GetAttendee operation.

See GetAttendee for more information about the operation and its arguments.

pub fn get_bot(&self) -> GetBot<C, M, R>

Constructs a fluent builder for the GetBot operation.

See GetBot for more information about the operation and its arguments.

pub fn get_channel_message(&self) -> GetChannelMessage<C, M, R>

Constructs a fluent builder for the GetChannelMessage operation.

See GetChannelMessage for more information about the operation and its arguments.

pub fn get_events_configuration(&self) -> GetEventsConfiguration<C, M, R>

Constructs a fluent builder for the GetEventsConfiguration operation.

See GetEventsConfiguration for more information about the operation and its arguments.

pub fn get_global_settings(&self) -> GetGlobalSettings<C, M, R>

Constructs a fluent builder for the GetGlobalSettings operation.

See GetGlobalSettings for more information about the operation and its arguments.

pub fn get_media_capture_pipeline(&self) -> GetMediaCapturePipeline<C, M, R>

Constructs a fluent builder for the GetMediaCapturePipeline operation.

See GetMediaCapturePipeline for more information about the operation and its arguments.

pub fn get_meeting(&self) -> GetMeeting<C, M, R>

Constructs a fluent builder for the GetMeeting operation.

See GetMeeting for more information about the operation and its arguments.

pub fn get_messaging_session_endpoint(
&self
) -> GetMessagingSessionEndpoint<C, M, R>

Constructs a fluent builder for the GetMessagingSessionEndpoint operation.

See GetMessagingSessionEndpoint for more information about the operation and its arguments.

pub fn get_phone_number(&self) -> GetPhoneNumber<C, M, R>

Constructs a fluent builder for the GetPhoneNumber operation.

See GetPhoneNumber for more information about the operation and its arguments.

pub fn get_phone_number_order(&self) -> GetPhoneNumberOrder<C, M, R>

Constructs a fluent builder for the GetPhoneNumberOrder operation.

See GetPhoneNumberOrder for more information about the operation and its arguments.

pub fn get_phone_number_settings(&self) -> GetPhoneNumberSettings<C, M, R>

Constructs a fluent builder for the GetPhoneNumberSettings operation.

See GetPhoneNumberSettings for more information about the operation and its arguments.

pub fn get_proxy_session(&self) -> GetProxySession<C, M, R>

Constructs a fluent builder for the GetProxySession operation.

See GetProxySession for more information about the operation and its arguments.

pub fn get_retention_settings(&self) -> GetRetentionSettings<C, M, R>

Constructs a fluent builder for the GetRetentionSettings operation.

See GetRetentionSettings for more information about the operation and its arguments.

pub fn get_room(&self) -> GetRoom<C, M, R>

Constructs a fluent builder for the GetRoom operation.

See GetRoom for more information about the operation and its arguments.

pub fn get_sip_media_application(&self) -> GetSipMediaApplication<C, M, R>

Constructs a fluent builder for the GetSipMediaApplication operation.

See GetSipMediaApplication for more information about the operation and its arguments.

pub fn get_sip_media_application_logging_configuration(
&self
) -> GetSipMediaApplicationLoggingConfiguration<C, M, R>

Constructs a fluent builder for the GetSipMediaApplicationLoggingConfiguration operation.

See GetSipMediaApplicationLoggingConfiguration for more information about the operation and its arguments.

pub fn get_sip_rule(&self) -> GetSipRule<C, M, R>

Constructs a fluent builder for the GetSipRule operation.

See GetSipRule for more information about the operation and its arguments.

pub fn get_user(&self) -> GetUser<C, M, R>

Constructs a fluent builder for the GetUser operation.

See GetUser for more information about the operation and its arguments.

pub fn get_user_settings(&self) -> GetUserSettings<C, M, R>

Constructs a fluent builder for the GetUserSettings operation.

See GetUserSettings for more information about the operation and its arguments.

pub fn get_voice_connector(&self) -> GetVoiceConnector<C, M, R>

Constructs a fluent builder for the GetVoiceConnector operation.

See GetVoiceConnector for more information about the operation and its arguments.

pub fn get_voice_connector_emergency_calling_configuration(
&self
) -> GetVoiceConnectorEmergencyCallingConfiguration<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorEmergencyCallingConfiguration operation.

See GetVoiceConnectorEmergencyCallingConfiguration for more information about the operation and its arguments.

pub fn get_voice_connector_group(&self) -> GetVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorGroup operation.

See GetVoiceConnectorGroup for more information about the operation and its arguments.

pub fn get_voice_connector_logging_configuration(
&self
) -> GetVoiceConnectorLoggingConfiguration<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorLoggingConfiguration operation.

See GetVoiceConnectorLoggingConfiguration for more information about the operation and its arguments.

pub fn get_voice_connector_origination(
&self
) -> GetVoiceConnectorOrigination<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorOrigination operation.

See GetVoiceConnectorOrigination for more information about the operation and its arguments.

pub fn get_voice_connector_proxy(&self) -> GetVoiceConnectorProxy<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorProxy operation.

See GetVoiceConnectorProxy for more information about the operation and its arguments.

pub fn get_voice_connector_streaming_configuration(
&self
) -> GetVoiceConnectorStreamingConfiguration<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorStreamingConfiguration operation.

See GetVoiceConnectorStreamingConfiguration for more information about the operation and its arguments.

pub fn get_voice_connector_termination(
&self
) -> GetVoiceConnectorTermination<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorTermination operation.

See GetVoiceConnectorTermination for more information about the operation and its arguments.

pub fn get_voice_connector_termination_health(
&self
) -> GetVoiceConnectorTerminationHealth<C, M, R>

Constructs a fluent builder for the GetVoiceConnectorTerminationHealth operation.

See GetVoiceConnectorTerminationHealth for more information about the operation and its arguments.

pub fn invite_users(&self) -> InviteUsers<C, M, R>

Constructs a fluent builder for the InviteUsers operation.

See InviteUsers for more information about the operation and its arguments.

pub fn list_accounts(&self) -> ListAccounts<C, M, R>

Constructs a fluent builder for the ListAccounts operation.

See ListAccounts for more information about the operation and its arguments.

pub fn list_app_instance_admins(&self) -> ListAppInstanceAdmins<C, M, R>

Constructs a fluent builder for the ListAppInstanceAdmins operation.

See ListAppInstanceAdmins for more information about the operation and its arguments.

pub fn list_app_instances(&self) -> ListAppInstances<C, M, R>

Constructs a fluent builder for the ListAppInstances operation.

See ListAppInstances for more information about the operation and its arguments.

pub fn list_app_instance_users(&self) -> ListAppInstanceUsers<C, M, R>

Constructs a fluent builder for the ListAppInstanceUsers operation.

See ListAppInstanceUsers for more information about the operation and its arguments.

pub fn list_attendees(&self) -> ListAttendees<C, M, R>

Constructs a fluent builder for the ListAttendees operation.

See ListAttendees for more information about the operation and its arguments.

pub fn list_attendee_tags(&self) -> ListAttendeeTags<C, M, R>

Constructs a fluent builder for the ListAttendeeTags operation.

See ListAttendeeTags for more information about the operation and its arguments.

pub fn list_bots(&self) -> ListBots<C, M, R>

Constructs a fluent builder for the ListBots operation.

See ListBots for more information about the operation and its arguments.

pub fn list_channel_bans(&self) -> ListChannelBans<C, M, R>

Constructs a fluent builder for the ListChannelBans operation.

See ListChannelBans for more information about the operation and its arguments.

pub fn list_channel_memberships(&self) -> ListChannelMemberships<C, M, R>

Constructs a fluent builder for the ListChannelMemberships operation.

See ListChannelMemberships for more information about the operation and its arguments.

pub fn list_channel_memberships_for_app_instance_user(
&self
) -> ListChannelMembershipsForAppInstanceUser<C, M, R>

Constructs a fluent builder for the ListChannelMembershipsForAppInstanceUser operation.

See ListChannelMembershipsForAppInstanceUser for more information about the operation and its arguments.

pub fn list_channel_messages(&self) -> ListChannelMessages<C, M, R>

Constructs a fluent builder for the ListChannelMessages operation.

See ListChannelMessages for more information about the operation and its arguments.

pub fn list_channel_moderators(&self) -> ListChannelModerators<C, M, R>

Constructs a fluent builder for the ListChannelModerators operation.

See ListChannelModerators for more information about the operation and its arguments.

pub fn list_channels(&self) -> ListChannels<C, M, R>

Constructs a fluent builder for the ListChannels operation.

See ListChannels for more information about the operation and its arguments.

pub fn list_channels_moderated_by_app_instance_user(
&self
) -> ListChannelsModeratedByAppInstanceUser<C, M, R>

Constructs a fluent builder for the ListChannelsModeratedByAppInstanceUser operation.

See ListChannelsModeratedByAppInstanceUser for more information about the operation and its arguments.

pub fn list_media_capture_pipelines(&self) -> ListMediaCapturePipelines<C, M, R>

Constructs a fluent builder for the ListMediaCapturePipelines operation.

See ListMediaCapturePipelines for more information about the operation and its arguments.

pub fn list_meetings(&self) -> ListMeetings<C, M, R>

Constructs a fluent builder for the ListMeetings operation.

See ListMeetings for more information about the operation and its arguments.

pub fn list_meeting_tags(&self) -> ListMeetingTags<C, M, R>

Constructs a fluent builder for the ListMeetingTags operation.

See ListMeetingTags for more information about the operation and its arguments.

pub fn list_phone_number_orders(&self) -> ListPhoneNumberOrders<C, M, R>

Constructs a fluent builder for the ListPhoneNumberOrders operation.

See ListPhoneNumberOrders for more information about the operation and its arguments.

pub fn list_phone_numbers(&self) -> ListPhoneNumbers<C, M, R>

Constructs a fluent builder for the ListPhoneNumbers operation.

See ListPhoneNumbers for more information about the operation and its arguments.

pub fn list_proxy_sessions(&self) -> ListProxySessions<C, M, R>

Constructs a fluent builder for the ListProxySessions operation.

See ListProxySessions for more information about the operation and its arguments.

pub fn list_room_memberships(&self) -> ListRoomMemberships<C, M, R>

Constructs a fluent builder for the ListRoomMemberships operation.

See ListRoomMemberships for more information about the operation and its arguments.

pub fn list_rooms(&self) -> ListRooms<C, M, R>

Constructs a fluent builder for the ListRooms operation.

See ListRooms for more information about the operation and its arguments.

pub fn list_sip_media_applications(&self) -> ListSipMediaApplications<C, M, R>

Constructs a fluent builder for the ListSipMediaApplications operation.

See ListSipMediaApplications for more information about the operation and its arguments.

pub fn list_sip_rules(&self) -> ListSipRules<C, M, R>

Constructs a fluent builder for the ListSipRules operation.

See ListSipRules for more information about the operation and its arguments.

pub fn list_supported_phone_number_countries(
&self
) -> ListSupportedPhoneNumberCountries<C, M, R>

Constructs a fluent builder for the ListSupportedPhoneNumberCountries operation.

See ListSupportedPhoneNumberCountries for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_users(&self) -> ListUsers<C, M, R>

Constructs a fluent builder for the ListUsers operation.

See ListUsers for more information about the operation and its arguments.

pub fn list_voice_connector_groups(&self) -> ListVoiceConnectorGroups<C, M, R>

Constructs a fluent builder for the ListVoiceConnectorGroups operation.

See ListVoiceConnectorGroups for more information about the operation and its arguments.

pub fn list_voice_connectors(&self) -> ListVoiceConnectors<C, M, R>

Constructs a fluent builder for the ListVoiceConnectors operation.

See ListVoiceConnectors for more information about the operation and its arguments.

pub fn list_voice_connector_termination_credentials(
&self
) -> ListVoiceConnectorTerminationCredentials<C, M, R>

Constructs a fluent builder for the ListVoiceConnectorTerminationCredentials operation.

See ListVoiceConnectorTerminationCredentials for more information about the operation and its arguments.

pub fn logout_user(&self) -> LogoutUser<C, M, R>

Constructs a fluent builder for the LogoutUser operation.

See LogoutUser for more information about the operation and its arguments.

pub fn put_app_instance_retention_settings(
&self
) -> PutAppInstanceRetentionSettings<C, M, R>

Constructs a fluent builder for the PutAppInstanceRetentionSettings operation.

See PutAppInstanceRetentionSettings for more information about the operation and its arguments.

pub fn put_app_instance_streaming_configurations(
&self
) -> PutAppInstanceStreamingConfigurations<C, M, R>

Constructs a fluent builder for the PutAppInstanceStreamingConfigurations operation.

See PutAppInstanceStreamingConfigurations for more information about the operation and its arguments.

pub fn put_events_configuration(&self) -> PutEventsConfiguration<C, M, R>

Constructs a fluent builder for the PutEventsConfiguration operation.

See PutEventsConfiguration for more information about the operation and its arguments.

pub fn put_retention_settings(&self) -> PutRetentionSettings<C, M, R>

Constructs a fluent builder for the PutRetentionSettings operation.

See PutRetentionSettings for more information about the operation and its arguments.

pub fn put_sip_media_application_logging_configuration(
&self
) -> PutSipMediaApplicationLoggingConfiguration<C, M, R>

Constructs a fluent builder for the PutSipMediaApplicationLoggingConfiguration operation.

See PutSipMediaApplicationLoggingConfiguration for more information about the operation and its arguments.

pub fn put_voice_connector_emergency_calling_configuration(
&self
) -> PutVoiceConnectorEmergencyCallingConfiguration<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorEmergencyCallingConfiguration operation.

See PutVoiceConnectorEmergencyCallingConfiguration for more information about the operation and its arguments.

pub fn put_voice_connector_logging_configuration(
&self
) -> PutVoiceConnectorLoggingConfiguration<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorLoggingConfiguration operation.

See PutVoiceConnectorLoggingConfiguration for more information about the operation and its arguments.

pub fn put_voice_connector_origination(
&self
) -> PutVoiceConnectorOrigination<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorOrigination operation.

See PutVoiceConnectorOrigination for more information about the operation and its arguments.

pub fn put_voice_connector_proxy(&self) -> PutVoiceConnectorProxy<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorProxy operation.

See PutVoiceConnectorProxy for more information about the operation and its arguments.

pub fn put_voice_connector_streaming_configuration(
&self
) -> PutVoiceConnectorStreamingConfiguration<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorStreamingConfiguration operation.

See PutVoiceConnectorStreamingConfiguration for more information about the operation and its arguments.

pub fn put_voice_connector_termination(
&self
) -> PutVoiceConnectorTermination<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorTermination operation.

See PutVoiceConnectorTermination for more information about the operation and its arguments.

pub fn put_voice_connector_termination_credentials(
&self
) -> PutVoiceConnectorTerminationCredentials<C, M, R>

Constructs a fluent builder for the PutVoiceConnectorTerminationCredentials operation.

See PutVoiceConnectorTerminationCredentials for more information about the operation and its arguments.

pub fn redact_channel_message(&self) -> RedactChannelMessage<C, M, R>

Constructs a fluent builder for the RedactChannelMessage operation.

See RedactChannelMessage for more information about the operation and its arguments.

pub fn redact_conversation_message(&self) -> RedactConversationMessage<C, M, R>

Constructs a fluent builder for the RedactConversationMessage operation.

See RedactConversationMessage for more information about the operation and its arguments.

pub fn redact_room_message(&self) -> RedactRoomMessage<C, M, R>

Constructs a fluent builder for the RedactRoomMessage operation.

See RedactRoomMessage for more information about the operation and its arguments.

pub fn regenerate_security_token(&self) -> RegenerateSecurityToken<C, M, R>

Constructs a fluent builder for the RegenerateSecurityToken operation.

See RegenerateSecurityToken for more information about the operation and its arguments.

pub fn reset_personal_pin(&self) -> ResetPersonalPIN<C, M, R>

Constructs a fluent builder for the ResetPersonalPIN operation.

See ResetPersonalPIN for more information about the operation and its arguments.

pub fn restore_phone_number(&self) -> RestorePhoneNumber<C, M, R>

Constructs a fluent builder for the RestorePhoneNumber operation.

See RestorePhoneNumber for more information about the operation and its arguments.

pub fn search_available_phone_numbers(
&self
) -> SearchAvailablePhoneNumbers<C, M, R>

Constructs a fluent builder for the SearchAvailablePhoneNumbers operation.

See SearchAvailablePhoneNumbers for more information about the operation and its arguments.

pub fn send_channel_message(&self) -> SendChannelMessage<C, M, R>

Constructs a fluent builder for the SendChannelMessage operation.

See SendChannelMessage for more information about the operation and its arguments.

pub fn start_meeting_transcription(&self) -> StartMeetingTranscription<C, M, R>

Constructs a fluent builder for the StartMeetingTranscription operation.

See StartMeetingTranscription for more information about the operation and its arguments.

pub fn stop_meeting_transcription(&self) -> StopMeetingTranscription<C, M, R>

Constructs a fluent builder for the StopMeetingTranscription operation.

See StopMeetingTranscription for more information about the operation and its arguments.

pub fn tag_attendee(&self) -> TagAttendee<C, M, R>

Constructs a fluent builder for the TagAttendee operation.

See TagAttendee for more information about the operation and its arguments.

pub fn tag_meeting(&self) -> TagMeeting<C, M, R>

Constructs a fluent builder for the TagMeeting operation.

See TagMeeting for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_attendee(&self) -> UntagAttendee<C, M, R>

Constructs a fluent builder for the UntagAttendee operation.

See UntagAttendee for more information about the operation and its arguments.

pub fn untag_meeting(&self) -> UntagMeeting<C, M, R>

Constructs a fluent builder for the UntagMeeting operation.

See UntagMeeting for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_account(&self) -> UpdateAccount<C, M, R>

Constructs a fluent builder for the UpdateAccount operation.

See UpdateAccount for more information about the operation and its arguments.

pub fn update_account_settings(&self) -> UpdateAccountSettings<C, M, R>

Constructs a fluent builder for the UpdateAccountSettings operation.

See UpdateAccountSettings for more information about the operation and its arguments.

pub fn update_app_instance(&self) -> UpdateAppInstance<C, M, R>

Constructs a fluent builder for the UpdateAppInstance operation.

See UpdateAppInstance for more information about the operation and its arguments.

pub fn update_app_instance_user(&self) -> UpdateAppInstanceUser<C, M, R>

Constructs a fluent builder for the UpdateAppInstanceUser operation.

See UpdateAppInstanceUser for more information about the operation and its arguments.

pub fn update_bot(&self) -> UpdateBot<C, M, R>

Constructs a fluent builder for the UpdateBot operation.

See UpdateBot for more information about the operation and its arguments.

pub fn update_channel(&self) -> UpdateChannel<C, M, R>

Constructs a fluent builder for the UpdateChannel operation.

See UpdateChannel for more information about the operation and its arguments.

pub fn update_channel_message(&self) -> UpdateChannelMessage<C, M, R>

Constructs a fluent builder for the UpdateChannelMessage operation.

See UpdateChannelMessage for more information about the operation and its arguments.

pub fn update_channel_read_marker(&self) -> UpdateChannelReadMarker<C, M, R>

Constructs a fluent builder for the UpdateChannelReadMarker operation.

See UpdateChannelReadMarker for more information about the operation and its arguments.

pub fn update_global_settings(&self) -> UpdateGlobalSettings<C, M, R>

Constructs a fluent builder for the UpdateGlobalSettings operation.

See UpdateGlobalSettings for more information about the operation and its arguments.

pub fn update_phone_number(&self) -> UpdatePhoneNumber<C, M, R>

Constructs a fluent builder for the UpdatePhoneNumber operation.

See UpdatePhoneNumber for more information about the operation and its arguments.

pub fn update_phone_number_settings(&self) -> UpdatePhoneNumberSettings<C, M, R>

Constructs a fluent builder for the UpdatePhoneNumberSettings operation.

See UpdatePhoneNumberSettings for more information about the operation and its arguments.

pub fn update_proxy_session(&self) -> UpdateProxySession<C, M, R>

Constructs a fluent builder for the UpdateProxySession operation.

See UpdateProxySession for more information about the operation and its arguments.

pub fn update_room(&self) -> UpdateRoom<C, M, R>

Constructs a fluent builder for the UpdateRoom operation.

See UpdateRoom for more information about the operation and its arguments.

pub fn update_room_membership(&self) -> UpdateRoomMembership<C, M, R>

Constructs a fluent builder for the UpdateRoomMembership operation.

See UpdateRoomMembership for more information about the operation and its arguments.

pub fn update_sip_media_application(&self) -> UpdateSipMediaApplication<C, M, R>

Constructs a fluent builder for the UpdateSipMediaApplication operation.

See UpdateSipMediaApplication for more information about the operation and its arguments.

pub fn update_sip_media_application_call(
&self
) -> UpdateSipMediaApplicationCall<C, M, R>

Constructs a fluent builder for the UpdateSipMediaApplicationCall operation.

See UpdateSipMediaApplicationCall for more information about the operation and its arguments.

pub fn update_sip_rule(&self) -> UpdateSipRule<C, M, R>

Constructs a fluent builder for the UpdateSipRule operation.

See UpdateSipRule for more information about the operation and its arguments.

pub fn update_user(&self) -> UpdateUser<C, M, R>

Constructs a fluent builder for the UpdateUser operation.

See UpdateUser for more information about the operation and its arguments.

pub fn update_user_settings(&self) -> UpdateUserSettings<C, M, R>

Constructs a fluent builder for the UpdateUserSettings operation.

See UpdateUserSettings for more information about the operation and its arguments.

pub fn update_voice_connector(&self) -> UpdateVoiceConnector<C, M, R>

Constructs a fluent builder for the UpdateVoiceConnector operation.

See UpdateVoiceConnector for more information about the operation and its arguments.

pub fn update_voice_connector_group(&self) -> UpdateVoiceConnectorGroup<C, M, R>

Constructs a fluent builder for the UpdateVoiceConnectorGroup operation.

See UpdateVoiceConnectorGroup for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.